% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven.R
\name{as.path}
\alias{as.path}
\alias{as.path.character}
\alias{as.path.module_definition}
\title{Generic to resove the path a path.}
\usage{
as.path(what, ...)

\method{as.path}{module_definition}(what, repositories, ...)

\method{as.path}{character}(what, ...)
}
\arguments{
\item{what}{The definition to resolve.}

\item{...}{Additional arguments passed to methods.}

\item{repositories}{A list of repositories to search in.}
}
\description{
Generic to resove the path a path.
}
\section{Methods (by class)}{
\itemize{
\item \code{module_definition}: Try to resolve a module in one of the given repositories, install it, and return the local path.

\item \code{character}: Simple returns the input
}}

