% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_from}
\alias{convert_from}
\title{Convert from transfer type to client type}
\usage{
convert_from(mapping, data, sql_type)
}
\arguments{
\item{a}{list of RJDBCMapping mapping definitions}

\item{data}{the data object to convert}

\item{sql_type}{the source sql type of the data}
}
\description{
Convert from transfer type to client type
}

