% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_mapping.R
\name{convert_from_transfer}
\alias{convert_from_transfer}
\title{Convert from transfer type to client type}
\usage{
convert_from_transfer(read_conversion_rules, data, data_attributes)
}
\arguments{
\item{read_conversion_rules}{a list of \code{\link{read_conversion_rule}} objects}

\item{data}{the data vector to convert}

\item{data_attributes}{a named list of attributes of data}
}
\description{
Convert from transfer type to client type
}
\keyword{internal}

