% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_to}
\alias{convert_to}
\title{Convert from client type to transfer type}
\usage{
convert_to(mapping, data, sql_type)
}
\arguments{
\item{a}{list of RJDBCMapping mapping definitions}

\item{data}{the data object to convert}

\item{sql_type}{the target sql type of the data}
}
\description{
Convert from client type to transfer type
}

