% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCMapping.R
\name{convert_to_transfer}
\alias{convert_to_transfer}
\title{Convert from client type to transfer type}
\usage{
convert_to_transfer(write_conversions, data, data_attributes)
}
\arguments{
\item{write_conversions}{a list of JDBCWriteConversion objects}

\item{data}{the data object to convert}

\item{data_attributes}{a named list of attributes}
}
\description{
Convert from client type to transfer type
}
\keyword{internal}

