% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_jdbc_utils.R
\name{create_j_colum}
\alias{create_j_colum}
\title{Create a Java Column class for given column_data}
\usage{
create_j_colum(column_data, sql_type, is_nullable, write_conversions)
}
\arguments{
\item{column_data}{the data to insert}

\item{sql_type}{the type of the column}

\item{is_nullable}{is the column nullable? 0 = disallows NULL, 1 = allows NULL, 2 = unknown}

\item{write_conversions}{a list of JDBCWriteConversion objects}
}
\description{
Create a Java Column class for given column_data
}

