% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_jdbc_utils.R
\name{create_jdbc_driver}
\alias{create_jdbc_driver}
\title{Create a Java JDBC Driver object}
\usage{
create_jdbc_driver(driverClass, classPath)
}
\arguments{
\item{driverClass}{a character vector of length one specifying a JDBC driver class (e.g. 'org.h2.Driver')}

\item{classPath}{a character vector of length one specifying classpaths separated by \code{\link[=.Platform]{path.sep}}
or a character vector of classpaths which will be added to the \code{\link[=.jaddClassPath]{rJava class loader}}}
}
\description{
Create a Java JDBC Driver object
}
\keyword{internal}

