% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_resultpull.R
\name{create_result_pull}
\alias{create_result_pull}
\title{Create a new JDBCjava_resultpull.Reference object}
\usage{
create_result_pull(j_result_set)
}
\arguments{
\item{j_result_set}{a jobjRef object to a java.sql.ResultSet object}
}
\description{
Create a new JDBCjava_resultpull.Reference object
}
\keyword{internal}

