% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbBegin,JDBCConnection-method}
\alias{dbBegin,JDBCConnection-method}
\title{Begin a transaction.}
\usage{
\S4method{dbBegin}{JDBCConnection}(conn, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Begin a transaction.
}

