% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbCallProc,JDBCConnection-method}
\alias{dbCallProc,JDBCConnection-method}
\title{Calls a stored procedure.}
\usage{
\S4method{dbCallProc}{JDBCConnection}(conn, name, parameters = list(), ...)
}
\arguments{
\item{conn}{a \code{\linkS4class{JDBCConnection}} object}

\item{name}{the name of the procedure}

\item{parameters}{a list of procedure arguments.}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Calls a stored procedure.
}

