% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R
\docType{methods}
\name{dbClearResult,JDBCQueryResult-method}
\alias{dbClearResult,JDBCQueryResult-method}
\title{Clear a result set.}
\usage{
\S4method{dbClearResult}{JDBCQueryResult}(res, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
Clear a result set.
}

