\docType{methods}
\name{dbClearResult,JDBCResult-method}
\alias{dbClearResult,JDBCResult-method}
\title{Clear a result set.}
\usage{
\S4method{dbClearResult}{JDBCResult}(res, ...)
}
\arguments{
  \item{res}{an \code{\linkS4class{JDBCResult}} object.}

  \item{...}{Ignored. Needed for compatibility with
  generic.}
}
\description{
Clear a result set.
}

