% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbColumnInfo,JDBCQueryResult-method}
\alias{dbColumnInfo,JDBCQueryResult-method}
\title{Get info about the result set data types.}
\usage{
\S4method{dbColumnInfo}{JDBCQueryResult}(res, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get info about the result set data types.
}

