\docType{methods}
\name{dbColumnInfo,JDBCResult-method}
\alias{dbColumnInfo,JDBCResult-method}
\title{Get info about the result set data types.}
\usage{
\S4method{dbColumnInfo}{JDBCResult}(res, ...)
}
\arguments{
  \item{res}{an \code{\linkS4class{JDBCResult}} object.}

  \item{...}{Ignored. Needed for compatiblity with
  generic.}
}
\description{
Get info about the result set data types.
}

