\docType{methods}
\name{dbConnect,JDBCConnection-method}
\alias{dbConnect,JDBCConnection-method}
\title{Create a JDBC connection.}
\usage{
\S4method{dbConnect}{JDBCConnection}(drv, ...)
}
\arguments{
  \item{drv}{a JDBCConnection object}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
Create a JDBC connection.
}

