% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCObject.R
\docType{methods}
\name{dbDataType,JDBCObject-method}
\alias{dbDataType,JDBCObject-method}
\title{Determine the SQL Data Type of an R object.}
\usage{
\S4method{dbDataType}{JDBCObject}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{a \code{\linkS4class{JDBCObject}} object.}

\item{obj}{an R object whose SQL type we want to determine.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
Determine the SQL Data Type of an R object.
}

