\docType{methods}
\name{dbGetException,JDBCConnection-method}
\alias{dbGetException,JDBCConnection-method}
\title{Get the last exception from the connection.}
\usage{
\S4method{dbGetException}{JDBCConnection}(conn, ...)
}
\arguments{
  \item{conn}{an object of class
  \code{\linkS4class{JDBCConnection}}}

  \item{...}{Ignored. Needed for compatiblity with
  generic.}
}
\description{
Get the last exception from the connection.
}

