\name{dbGetFields}
\alias{dbGetFields}
\title{Generics for getting description of table columns available in the specified catalog.}
\usage{
dbGetFields(conn, ...)
}
\arguments{
  \item{conn}{An \code{DBIConnection} object.}

  \item{...}{Other arguments used by methods}
}
\description{
Generics for getting description of table columns available
in the specified catalog.
}

