\docType{methods}
\name{dbGetInfo,JDBCObject-method}
\alias{dbGetInfo,JDBCObject-method}
\title{Get metadata about a database object.}
\usage{
\S4method{dbGetInfo}{JDBCObject}(dbObj, ...)
}
\arguments{
  \item{dbObj}{An object of class
  \code{\linkS4class{JDBCObject}}}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
Get metadata about a database object.
}

