% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbGetInfo,JDBCQueryResult-method}
\alias{dbGetInfo,JDBCQueryResult-method}
\title{Get info about the result.}
\usage{
\S4method{dbGetInfo}{JDBCQueryResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of class \code{\linkS4class{JDBCQueryResult}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get info about the result.
}

