\docType{methods}
\name{dbGetInfo,JDBCResult-method}
\alias{dbGetInfo,JDBCResult-method}
\title{Get info about the result.}
\usage{
\S4method{dbGetInfo}{JDBCResult}(dbObj, ...)
}
\arguments{
  \item{dbObj}{an object of class
  \code{\linkS4class{JDBCResult}}}

  \item{...}{Ignored. Needed for compatiblity with
  generic.}
}
\description{
Get info about the result.
}

