% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbGetInfo,JDBCUpdateResult-method}
\alias{dbGetInfo,JDBCUpdateResult-method}
\title{Get info about the result.}
\usage{
\S4method{dbGetInfo}{JDBCUpdateResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of class \code{\linkS4class{JDBCUpdateResult}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get info about the result.
}

