% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbGetRowCount,JDBCUpdateResult-method}
\alias{dbGetRowCount,JDBCUpdateResult-method}
\title{Get number of rows fetched so far}
\usage{
\S4method{dbGetRowCount}{JDBCUpdateResult}(res, ...)
}
\description{
Get number of rows fetched so far
}

