% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection_generics.R
\name{dbGetTables}
\alias{dbGetTables}
\title{Generics for getting a description of the tables available in the given catalog.}
\usage{
dbGetTables(conn, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object.}

\item{...}{Other arguments used by methods}
}
\description{
Generics for getting a description of the tables available in the given catalog.
}
\keyword{internal}

