% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbIsValid,JDBCConnection-method}
\alias{dbIsValid,JDBCConnection-method}
\title{dbIsValid}
\usage{
\S4method{dbIsValid}{JDBCConnection}(dbObj, timeout = 0, ...)
}
\arguments{
\item{dbObj}{A subclass of DBIConnection, representing an active connection to an DBMS.}

\item{timeout}{The time in seconds to wait for the database operation used to validate the connection to complete.
If the timeout period expires before the operation completes, this method returns false.
A value of 0 indicates a timeout is not applied to the database operation.}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
dbIsValid
}

