% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{methods}
\name{dbIsValid,JDBCDriver-method}
\alias{dbIsValid,JDBCDriver-method}
\title{Check if \code{dbObj} is valid, which is always true for objects of type \code{\linkS4class{JDBCDriver}}.}
\usage{
\S4method{dbIsValid}{JDBCDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of class \code{\linkS4class{JDBCDriver}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Check if \code{dbObj} is valid, which is always true for objects of type \code{\linkS4class{JDBCDriver}}.
}

