% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R
\docType{methods}
\name{dbIsValid,JDBCQueryResult-method}
\alias{dbIsValid,JDBCQueryResult-method}
\title{Get info about the result.}
\usage{
\S4method{dbIsValid}{JDBCQueryResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of class \code{\linkS4class{JDBCQueryResult}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get info about the result.
}

