% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCUpdateResult.R
\docType{methods}
\name{dbIsValid,JDBCUpdateResult-method}
\alias{dbIsValid,JDBCUpdateResult-method}
\title{Check if \code{dbObj} is valid, which is always true for objects of type \code{\linkS4class{JDBCUpdateResult}}.}
\usage{
\S4method{dbIsValid}{JDBCUpdateResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of class \code{\linkS4class{JDBCUpdateResult}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Check if \code{dbObj} is valid, which is always true for objects of type \code{\linkS4class{JDBCUpdateResult}}.
}

