% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbj-edit-tables}
\alias{dbReadTable,JDBCConnection,character-method}
\alias{dbTruncateTable,JDBCConnection,character-method}
\alias{dbWriteTable,JDBCConnection,character,data.frame-method}
\alias{dbWriteTable,JDBCConnection,character-data.frame-method}
\alias{dbj-edit-tables}
\title{Modify data in a database table}
\usage{
\S4method{dbReadTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbWriteTable}{JDBCConnection,character,data.frame}(conn, name, value,
  create = TRUE, append = FALSE, truncate = FALSE, temporary = FALSE,
  row.names = NA, sql.types = NULL, ...)

\S4method{dbTruncateTable}{JDBCConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{name}{character vector of length 1 specifying the name of a table in the database}

\item{...}{Other parameters passed on to methods.}

\item{value}{the date frame to write to the table}

\item{create}{a logical specifying whether to create a new table if it does not exist. Its default is TRUE.}

\item{append}{a logical specifying whether to append to an existing table. Its default is FALSE}

\item{truncate}{a logical specifying whether to truncate an existing table before appending. Its default is FALSE}

\item{temporary}{\code{TRUE} if the table should be temporary}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

  If \code{TRUE}, always translate row names to a column called "row_names".
  If \code{FALSE}, never translate row names. If \code{NA}, translate
  rownames only if they're a character vector.

  A string is equivalent to \code{TRUE}, but allows you to override the
  default name.

  For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{sql.types}{A character vector of named SQL field types where
the names are the names of new table's columns. If NULL, types will be inferred
with \code{\link[DBI]{dbDataType}}).}
}
\description{
Read data from a table, write a data frame to a table or delete the content of a table.
}
\details{
These methods use \code{\link{sql_dialect}} functions to generate SQL statments.
}
\examples{
\dontrun{
dbWriteTable(con, iris)
dbReadTable(con, "iris")
dbTruncateTable(con, "iris")
}
}
\seealso{
Other JDBCConnection methods: \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbDisconnect,JDBCConnection-method}},
  \code{\link{dbSendQuery,JDBCConnection,character-method}},
  \code{\link{dbj-sql-generate}},
  \code{\link{dbj-sql-quote}}, \code{\link{dbj-table}}
}

