% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbj-sql-generate}
\alias{dbj-sql-generate}
\alias{sqlAppendTable,JDBCConnection-method}
\alias{sqlCreateTable,JDBCConnection-method}
\title{Generate SQL statements}
\usage{
\S4method{sqlCreateTable}{JDBCConnection}(con, table, fields, row.names = NA,
  temporary = FALSE, ...)

\S4method{sqlAppendTable}{JDBCConnection}(con, table, values, row.names = NA,
  ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{fields}{Either a character vector or a data frame.

  A named character vector: Names are column names, values are types.
  Names are escaped with \code{\link[DBI]{dbQuoteIdentifier}}.
  Field types are unescaped.

  A data frame: field types are generated using
  \code{\link[DBI]{dbDataType}}.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

  If \code{TRUE}, always translate row names to a column called "row_names".
  If \code{FALSE}, never translate row names. If \code{NA}, translate
  rownames only if they're a character vector.

  A string is equivalent to \code{TRUE}, but allows you to override the
  default name.

  For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}

\item{...}{Other arguments used by individual methods.}

\item{values}{A data frame. Factors will be converted to character vectors.
Character vectors will be escaped with \code{\link[DBI]{dbQuoteString}}.}
}
\value{
an \code{\link{SQL}} object
}
\description{
These functions use the driver's \code{sql_dialect} environment
to generate SQL statments for creating tables and inserting data.
}
\seealso{
Other JDBCConnection methods: \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbDisconnect,JDBCConnection-method}},
  \code{\link{dbSendQuery,JDBCConnection,character-method}},
  \code{\link{dbj-edit-tables}},
  \code{\link{dbj-sql-quote}}, \code{\link{dbj-table}}

Other SQL dialect functions: \code{\link{dbj-sql-quote}},
  \code{\link{sql_dialect}}

Other SQL functions: \code{\link{dbj-sql-quote}}
}

