% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbj-sql-quote}
\alias{dbQuoteIdentifier,JDBCConnection,character-method}
\alias{dbQuoteString,JDBCConnection,character-method}
\alias{dbj-sql-quote}
\title{SQL quoting.}
\usage{
\S4method{dbQuoteIdentifier}{JDBCConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{JDBCConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{x}{A character vector to quote.}

\item{...}{Other parameters passed on to methods.}
}
\value{
an \code{\link{SQL}} object
}
\description{
These functions use the driver's \code{sql_dialect} environment to quote strings and idetifiers.
}
\seealso{
Other JDBCConnection methods: \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbDisconnect,JDBCConnection-method}},
  \code{\link{dbSendQuery,JDBCConnection,character-method}},
  \code{\link{dbj-edit-tables}},
  \code{\link{dbj-sql-generate}}, \code{\link{dbj-table}}

Other SQL dialect functions: \code{\link{dbj-sql-generate}},
  \code{\link{sql_dialect}}

Other SQL functions: \code{\link{dbj-sql-generate}}
}

