% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbj-table}
\alias{dbExistsTable,JDBCConnection,character-method}
\alias{dbExistsTable,JDBCConnection-method}
\alias{dbGetFields,JDBCConnection-method}
\alias{dbGetTables,JDBCConnection-method}
\alias{dbListFields,JDBCConnection,character-method}
\alias{dbListTables,JDBCConnection-method}
\alias{dbRemoveTable,JDBCConnection,character-method}
\alias{dbRemoveTable,JDBCConnection-method}
\alias{dbj-table}
\title{dbj table handling}
\usage{
\S4method{dbListTables}{JDBCConnection}(conn, pattern = "\%", ...)

\S4method{dbGetTables}{JDBCConnection}(conn, pattern = "\%", ...)

\S4method{dbExistsTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbListFields}{JDBCConnection,character}(conn, name, pattern = "\%",
  ...)

\S4method{dbGetFields}{JDBCConnection}(conn, name, pattern = "\%", ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{pattern}{the pattern to filter tables or fields}

\item{...}{Other parameters passed on to methods.}

\item{name}{A character vector of length one specifying a table name.}
}
\description{
dbj table handling
}
\seealso{
Other JDBCConnection methods: \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbDisconnect,JDBCConnection-method}},
  \code{\link{dbSendQuery,JDBCConnection,character-method}},
  \code{\link{dbj-edit-tables}},
  \code{\link{dbj-sql-generate}},
  \code{\link{dbj-sql-quote}}
}
\keyword{internal}

