% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbListFields,JDBCConnection,character-method}
\alias{dbListFields,JDBCConnection,character-method}
\title{List field names in specified table.}
\usage{
\S4method{dbListFields}{JDBCConnection,character}(conn, name, pattern = "\%",
  ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{name}{A character string specifying a DBMS table name.}

\item{pattern}{the pattern for the columns to list}

\item{...}{Other parameters passed on to methods.}
}
\description{
List field names in specified table.
}

