\docType{methods}
\name{dbListResults,JDBCConnection-method}
\alias{dbListResults,JDBCConnection-method}
\title{List available JDBC result sets.}
\usage{
\S4method{dbListResults}{JDBCConnection}(conn, ...)
}
\arguments{
  \item{conn}{An existing
  \code{\linkS4class{JDBCConnection}}}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
List available JDBC result sets.
}

