% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbListTables,JDBCConnection-method}
\alias{dbListTables,JDBCConnection-method}
\title{List available JDBC tables.}
\usage{
\S4method{dbListTables}{JDBCConnection}(conn, pattern = "\%", ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{pattern}{the pattern passed to the java method ResultSet.getTables()}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
List available JDBC tables.
}

