% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbQuoteIdentifier,JDBCConnection,character-method}
\alias{dbQuoteIdentifier,JDBCConnection,character-method}
\title{Quote identifier}
\usage{
\S4method{dbQuoteIdentifier}{JDBCConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{A subclass of DBIConnection, representing an active connection to an DBMS.}

\item{x}{A character vector to label as being escaped SQL.}

\item{...}{Other arguments passed on to methods. Not otherwise used.}
}
\description{
Quote identifier
}

