% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbRemoveTable,JDBCConnection,character-method}
\alias{dbRemoveTable,JDBCConnection,character-method}
\title{Remove a table from the database.}
\usage{
\S4method{dbRemoveTable}{JDBCConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{name}{character vector of length 1 giving name of table to remove}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Executes the SQL \code{DROP TABLE}.
}

