% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbRollback,JDBCConnection-method}
\alias{dbRollback,JDBCConnection-method}
\title{Roll back a transaction.}
\usage{
\S4method{dbRollback}{JDBCConnection}(conn, ...)
}
\arguments{
\item{conn}{an object of class \code{\linkS4class{JDBCConnection}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Roll back a transaction.
}

