% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbSendQuery,JDBCConnection,character-method}
\alias{dbSendQuery,JDBCConnection,character-method}
\title{Execute a SQL statement on a database connection}
\usage{
\S4method{dbSendQuery}{JDBCConnection,character}(conn, statement,
  parameters = list())
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{statement}{A SQL statement to send over the connection. Use \code{?} for input parameters.}

\item{parameters}{A list of statement parameters, which will be inserted in order.}
}
\value{
A \code{\linkS4class{JDBCResult}} object
}
\description{
Execute a SQL statement on a database connection
}

