% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbSendUpdate,JDBCConnection,character,data.frame-method}
\alias{dbSendUpdate,JDBCConnection,character,data.frame-method}
\title{Update a database using a prepared statment query and multiple parameters.}
\usage{
\S4method{dbSendUpdate}{JDBCConnection,character,data.frame}(conn, statement,
  parameters, partition_size = 10000, ...)
}
\arguments{
\item{partition_size}{the size which will be used to partition the data into seperate commits}

\item{conn}{An \code{DBIConnection} object.}

\item{statement}{the statement to send}

\item{parameters}{Optional. Either a named list or a data.frame of statment parameters.
A data.frame will produce a batch update for each row.}

\item{...}{Other arguments used by methods}
}
\description{
Update a database using a prepared statment query and multiple parameters.
}

