% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection_generics.R
\name{dbSQLDialect}
\alias{dbSQLDialect}
\title{Get the current SQL dialect of the connection}
\usage{
dbSQLDialect(conn, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object.}

\item{...}{Other arguments used by methods}
}
\description{
Get the current SQL dialect of the connection
}

