% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbTruncateTable,JDBCConnection,character-method}
\alias{dbTruncateTable,JDBCConnection,character-method}
\title{Truncate a table}
\usage{
\S4method{dbTruncateTable}{JDBCConnection,character}(conn, name,
  use_delete = FALSE, ...)
}
\arguments{
\item{conn}{An \code{DBIConnection} object.}

\item{name}{The name of the table.}

\item{use_delete}{Send a DELETE FROM query instead of TRUNCATE TABLE. Default is FALSE.}

\item{...}{Other arguments used by methods.}
}
\description{
Truncate a table
}

