% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnectionExtensions.R
\name{dbTruncateTable}
\alias{dbTruncateTable}
\title{Truncate a table}
\usage{
dbTruncateTable(conn, name, ...)
}
\arguments{
\item{conn}{An \code{DBIConnection} object.}

\item{name}{The name of the table.}

\item{...}{Other arguments used by methods.}
}
\description{
Truncate a table
}
\keyword{internal}

