\docType{methods}
\name{dbUnloadDriver,JDBCDriver-method}
\alias{dbUnloadDriver,JDBCDriver-method}
\title{Unload JDBCDriver driver.}
\usage{
\S4method{dbUnloadDriver}{JDBCDriver}(drv, ...)
}
\arguments{
  \item{drv}{An object of class
  \code{\linkS4class{JDBCDriver}}}

  \item{...}{Ignored. Needed for compatibility with
  generic.}
}
\value{
A logical indicating whether the operation succeeded or
not.
}
\description{
Unload JDBCDriver driver.
}

