% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbWriteTable,JDBCConnection,character,data.frame-method}
\alias{dbWriteTable,JDBCConnection,character,data.frame-method}
\title{Write a local data frame or file to the database.}
\usage{
\S4method{dbWriteTable}{JDBCConnection,character,data.frame}(conn, name, value,
  create = TRUE, append = TRUE, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{name}{character vector of length 1 giving name of table to write to}

\item{value}{the date frame to write to the table}

\item{create}{a logical specifying whether to create a new table if it does not exist. Its default is TRUE.}

\item{append}{a logical specifying whether to append to an existing table. Its default is TRUE.}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Write a local data frame or file to the database.
}

