% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_mapping_defaults.R
\docType{data}
\name{default-conversion-rules}
\alias{default-conversion-rules}
\alias{default_read_conversions}
\alias{default_write_conversions}
\title{Default type conversion rules}
\format{An object of class \code{list} of length 9.}
\usage{
default_read_conversions

default_write_conversions
}
\description{
Default type conversion rules
}
\keyword{datasets}

