% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\name{driver}
\alias{driver}
\title{Factory function for \code{\linkS4class{JDBCDriver}} objects}
\usage{
driver(driver_class, classpath = NULL,
  dialect = guess_dialect(driver_class),
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions,
  create_new_connection = create_new_dbj_connection)
}
\arguments{
\item{driver_class}{a character vector of length one specifying a JDBC driver class (e.g. 'org.h2.Driver')}

\item{classpath}{a character vector of length one specifying classpaths separated by \code{\link[=.Platform]{path.sep}}
or a character vector of classpaths which will be added to the \code{\link[=.jaddClassPath]{rJava class loader}}}

\item{dialect}{The \code{\link{sql_dialect}} to use.}

\item{read_conversions}{a list of \code{\link[=read_conversion_rule]{read conversions}}.}

\item{write_conversions}{a list of \code{\link[=write_conversion_rule]{write conversions}}.}

\item{create_new_connection}{The factory function for JDBCConnection objects.}
}
\value{
A new \code{\linkS4class{JDBCDriver}}
}
\description{
Call \code{driver} to create a new \code{\linkS4class{JDBCDriver}}
in order to \code{\link[=dbConnect,JDBCDriver-method]{connect}} to databases using the given JDBC driver.
}
\examples{
\dontrun{
drv <- dbj::driver('org.h2.Driver', '~/h2.jar')
con1 <- dbConnect(drv, 'jdbc:h2:mem:')
con2 <- dbConnect(drv, 'jdbc:h2:file:~/foo')

# Alternatively load the classpath directly with rJava
# using \\code{\\link[rJava]{.jinit}} or \\code{\\link[rJava]{.jpackage}} e.g.:
rJava::.jinit('~/h2.jar')
dbj::driver('org.h2.Driver')
}
}

