% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R
\docType{methods}
\name{fetch,JDBCQueryResult,numeric-method}
\alias{fetch,JDBCQueryResult,missing-method}
\alias{fetch,JDBCQueryResult,numeric-method}
\title{Fetch records from a previously executed query}
\usage{
\S4method{fetch}{JDBCQueryResult,numeric}(res, n, fetch_size = 0, ...)

\S4method{fetch}{JDBCQueryResult,missing}(res, n = -1, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{n}{optional maximum number of records to retrieve per fetch. Use \code{-1} to
retrieve all pending records; use \code{0} for to fetch the default
number of rows as defined in \code{\link{JDBC}}}

\item{fetch_size}{a hint to the number of rows that should be fetched from the database in a single block.
See \url{http://docs.oracle.com/javase/7/docs/api/java/sql/Statement.html#setFetchSize(int)}.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
Fetch records from a previously executed query
}

