% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_jdbc_utils.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Create a new ResultSetMetaData reference object}
\usage{
get_meta_data(j_result_set)
}
\arguments{
\item{j_result_set}{a jobjRef object to a java.sql.ResultSet object}
}
\description{
Create a new ResultSetMetaData reference object
}
\keyword{internal}

