% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCUtils.R
\name{insert_parameters}
\alias{insert_parameters}
\title{Set the values of prepared statment.}
\usage{
insert_parameters(j_statement, parameter_list, write_conversions)
}
\arguments{
\item{j_statement}{a java reference object to a java.sql.PreparedStatement}

\item{parameter_list}{a list of parameter values to fill the statement with}

\item{write_conversions}{a list of \code{JDBCWriteConversion} objects}
}
\description{
Set the values of prepared statment.
}

