% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven.R
\name{java_classpath}
\alias{java_classpath}
\title{Create a classpath from the arguments.
Character verctors are treated as local file paths.
\code{module} objects will be resolved.}
\usage{
java_classpath(..., repositories = list(maven_central))
}
\arguments{
\item{...}{Either local file paths or objects created with \code{module}.}

\item{repositories}{The repositories used to resolve modules.}
}
\description{
Create a classpath from the arguments.
Character verctors are treated as local file paths.
\code{module} objects will be resolved.
}

