% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_connect}
\alias{jdbc_connect}
\title{Establish a JDBC Connection}
\usage{
jdbc_connect(url, user, password, ...)
}
\arguments{
\item{url}{the URL of the form \code{jdbc:subprotocol:subname}}

\item{user}{the user to log in}

\item{password}{the user's password}

\item{...}{additional connection arguments}
}
\value{
a jObjRef referencing a java.sql.Connection
}
\description{
Establish a JDBC Connection
}
\keyword{internal}

