% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_create_driver}
\alias{jdbc_create_driver}
\title{Create a Java JDBC Driver object}
\usage{
jdbc_create_driver(driver_class, classpath = NULL)
}
\arguments{
\item{driver_class}{a character vector specifying the JDBC driver classes (e.g. 'org.h2.Driver')}

\item{classpath}{a character vector of length one specifying classpaths separated by \code{\link[=.Platform]{path.sep}}
or a character vector of classpaths which will be added to the \code{\link[=.jaddClassPath]{rJava class loader}}}
}
\description{
Create a Java JDBC Driver object
}
\keyword{internal}

