% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_create_resultpull}
\alias{jdbc_create_resultpull}
\title{Create a new JDBCjava_resultpull.Reference object}
\usage{
jdbc_create_resultpull(j_result_set)
}
\arguments{
\item{j_result_set}{a jobjRef object to a java.sql.ResultSet object}
}
\description{
Create a new JDBCjava_resultpull.Reference object
}
\keyword{internal}

