% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_create_table}
\alias{jdbc_create_table}
\title{Transform a data frame into a Java reference to a com/github/hoesler/dbj/Table}
\usage{
jdbc_create_table(data, write_conversions)
}
\arguments{
\item{data}{a data.frame}

\item{write_conversions}{a list of \code{\link{write_conversion_rule}} objects}
}
\description{
Transform a data frame into a Java reference to a com/github/hoesler/dbj/Table
}
\keyword{internal}

