% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_create_table_colum}
\alias{jdbc_create_table_colum}
\title{Create a Java Column class for given column_data}
\usage{
jdbc_create_table_colum(column_data, write_conversions)
}
\arguments{
\item{column_data}{the data to insert}

\item{write_conversions}{a list of \code{\link{write_conversion_rule}} objects}
}
\description{
Create a Java Column class for given column_data
}
\keyword{internal}

