% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_execute_batch}
\alias{jdbc_execute_batch}
\title{Execute a batch statement}
\usage{
jdbc_execute_batch(j_statement)
}
\arguments{
\item{j_statement}{a jobjRef object to a java.sql.PreparedStatement object}
}
\value{
integer vector with update counts containing one element for each command in the batch
}
\description{
Execute a batch statement
}
\keyword{internal}

