% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_jdbc_utils.R
\name{jdbc_register_driver}
\alias{jdbc_register_driver}
\title{Register a JDBC Driver}
\usage{
jdbc_register_driver(driver_class, classpath = NULL)
}
\arguments{
\item{driver_class}{a character vector specifying the JDBC driver classes (e.g. 'org.h2.Driver')}

\item{classpath}{a character vector of length one specifying classpaths separated by \code{\link[=.Platform]{path.sep}}
or a character vector of classpaths which will be added to the \code{\link[=.jaddClassPath]{rJava class loader}}}
}
\description{
Register a JDBC Driver
}

