% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_resultset_meta}
\alias{jdbc_resultset_meta}
\title{Create a new ResultSetMetaData reference object}
\usage{
jdbc_resultset_meta(j_result_set)
}
\arguments{
\item{j_result_set}{a jobjRef object to a java.sql.ResultSet object}
}
\description{
Create a new ResultSetMetaData reference object
}
\keyword{internal}

