% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{jdbc_set_statement_parameters}
\alias{jdbc_set_statement_parameters}
\title{Set the values of prepared statement.}
\usage{
jdbc_set_statement_parameters(j_statement, parameter_list, write_conversions)
}
\arguments{
\item{j_statement}{a Java reference object to a java.sql.PreparedStatement}

\item{parameter_list}{a list of parameter values to fill the statement with}

\item{write_conversions}{a list of \code{\link{write_conversion_rule}} objects}
}
\description{
Set the values of prepared statement.
}
\keyword{internal}

