% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_sql_types.R
\docType{data}
\name{JDBC_SQL_TYPES}
\alias{JDBC_SQL_TYPES}
\title{A named list of SQL types as defined in the java class
\href{https://docs.oracle.com/javase/7/docs/api/java/sql/Types.html}{java.sql.Types}}
\format{An object of class \code{list} of length 36.}
\usage{
JDBC_SQL_TYPES
}
\description{
A named list of SQL types as defined in the java class
\href{https://docs.oracle.com/javase/7/docs/api/java/sql/Types.html}{java.sql.Types}
}
\keyword{datasets}

