% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\name{JDBCConnection}
\alias{JDBCConnection}
\title{Create a new \code{\linkS4class{JDBCConnection}} object}
\usage{
JDBCConnection(j_connection, driver)
}
\arguments{
\item{j_connection}{a jobjRef object with a java.sql.Connection reference}

\item{driver}{the full qualified class name of the JDBC Driver to use.}
}
\value{
a \code{\linkS4class{JDBCConnection}} object
}
\description{
Create a new \code{\linkS4class{JDBCConnection}} object
}
\keyword{internal}

