% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\name{JDBC}
\alias{JDBC}
\title{Deprecated}
\usage{
JDBC(driverClass = "", classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions)
}
\arguments{
\item{driverClass}{the java class name of the JDBC driver to use}

\item{classPath}{a string of paths seperated by : which should get added to the classpath (see \link[rJava]{.jaddClassPath})}

\item{read_conversions}{a list of JDBCReadConversion objects.}

\item{write_conversions}{a list of JDBCWriteConversion objects.}
}
\description{
Deprecated
}

