% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\name{JDBC}
\alias{JDBC}
\title{Legacy JDBCDriver constructor.}
\usage{
JDBC(driverClass = "", classPath = "", ...)
}
\arguments{
\item{driverClass}{the java class name of the JDBC driver to use}

\item{classPath}{a string of paths seperated by the \code{path.sep} variable in \code{\link{.Platform}} which should get added to the classpath (see \link[rJava]{.jaddClassPath})}

\item{...}{Ignored.}
}
\description{
This function is present for \code{RJDBC} compatibility. Use \code{\link{driver}} instead.
}
\keyword{internal}

