% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCResult.R
\docType{class}
\name{JDBCResult-class}
\alias{JDBCResult-class}
\alias{SQLKeywords,JDBCResult-method}
\alias{dbGetDriver,JDBCResult-method}
\title{JDBCResult class.}
\usage{
\S4method{dbGetDriver}{JDBCResult}(dbObj, ...)

\S4method{SQLKeywords}{JDBCResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object of class \code{\linkS4class{JDBCResult}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Base class for \code{linkS4class{JDBCQueryResult}} and \code{linkS4class{JDBCUpdateResult}}.
}

