% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jstop}
\alias{jstop}
\title{Throw an error with the given \code{j_exception} as the cause.}
\usage{
jstop(j_exception, expression, ...)
}
\arguments{
\item{j_exception}{a java Throwable which is converted to a part of the message}

\item{expression}{a R expression to evaluate}

\item{...}{any further message parts passed to \code{stop}}
}
\description{
Throw an error with the given \code{j_exception} as the cause.
}

