% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven.R
\docType{data}
\name{maven}
\alias{default_maven_local}
\alias{maven}
\alias{maven_central}
\alias{maven_jar}
\title{Maven utils}
\format{An object of class \code{character} of length 1.}
\usage{
maven_central

default_maven_local

maven_jar(group_id, artifact_id, version,
  local_repository = default_maven_local, remote_repository = maven_central)
}
\arguments{
\item{group_id}{The group id of the artifact}

\item{artifact_id}{The id of the artifact}

\item{version}{The version of the artifact}

\item{local_repository}{The path to the local repository}

\item{remote_repository}{The url of the remote repository}
}
\value{
The path to the jar file
}
\description{
Maven utils

Get the path to the desired artifact in the \code{local_repository} and install from the \code{remote_repository} if missing.
}
\keyword{datasets}

