% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven.R
\docType{data}
\name{maven_repository}
\alias{maven_central}
\alias{maven_repository}
\title{Create a maven repository definition}
\format{An object of class \code{maven_repository} (inherits from \code{module_repository}) of length 1.}
\usage{
maven_repository(url)

maven_central
}
\arguments{
\item{url}{The url of the reopsitory.}
}
\description{
Create a maven repository definition
}
\keyword{datasets}

