% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_dependencies.R
\name{module}
\alias{module}
\alias{module.character}
\alias{module.list}
\title{Define Java dependencies as a module}
\usage{
module(x, ...)

\method{module}{character}(x, transitive = TRUE, ...)

\method{module}{list}(x, transitive = TRUE, ...)
}
\arguments{
\item{x}{A module definition}

\item{...}{Additional arguments passed to methods.}

\item{transitive}{A logical indicating if transitive dependencies should be fetched as well}
}
\value{
a list with class \code{"module"}
}
\description{
The \code{module} function creates a definition object of a Java dependency,
which can be \code{\link[=resolve]{resolved}} from a \code{\link{repository}}
and used to create a classpath argument to \code{\link{driver}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Constructs a module from a string in the form 'group:name:version'.

\item \code{list}: Constructs a module from a named list with elements \code{group}, \code{name} and \code{version}.
}}
\examples{
module('com.h2database:h2:1.3.176')
module('com.h2database:h2:1.3.176', transitive = FALSE)
module(list(group = 'com.h2database', name = 'h2', version = '1.3.176'))
}
\seealso{
Other java dependency functions: \code{\link{fetch_module}},
  \code{\link{maven_repository}}, \code{\link{resolve}}
}

