% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RUtils.R
\name{partition}
\alias{partition}
\alias{partition.data.frame}
\alias{partition.default}
\title{Partition a vector, list matrix or data.frame into a list of vectors, lists, matrices or data.frames respectively,
each of the same size (last might be smaller).}
\usage{
partition(data, size)

\method{partition}{data.frame}(data, size)

\method{partition}{default}(data, size)
}
\arguments{
\item{data}{the data to partition}

\item{size}{the partition size}
}
\description{
Partition a vector, list matrix or data.frame into a list of vectors, lists, matrices or data.frames respectively,
each of the same size (last might be smaller).
}

