% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_mapping.R
\name{read_conversion_rule}
\alias{read_conversion_rule}
\title{Create a read conversion rule}
\usage{
read_conversion_rule(condition, r_class, conversion)
}
\arguments{
\item{condition}{a function which accepts a list of data attributes and returns a logical
indicating if it is able to convert the data}

\item{r_class}{the target class of the conversion}

\item{conversion}{a function which accepts data of a transfer type
and transforms into data of type \code{r_class}.}
}
\description{
Read conversion rules define how to convert data read from JDBC to R values.
}
\examples{

# Convert TIME data, fetched from JDBC as numeric milliseconds, to difftime.
read_conversion_rule(
  function(jdbc.type, ...) with(JDBC_SQL_TYPES,
    jdbc.type == TIME),
  function(...) "difftime",
  function(data, ...) as.difftime(data / 1000, units = "secs")
)
}
\seealso{
Other conversion functions: \code{\link{write_conversion_rule}}
}

