% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_dependencies.R
\name{resolve}
\alias{resolve}
\alias{resolve.character}
\alias{resolve.list}
\alias{resolve.module}
\title{Resolve objects to file paths}
\usage{
resolve(what, ...)

\method{resolve}{module}(what, repositories, transitive = TRUE, ...)

\method{resolve}{character}(what, ...)

\method{resolve}{list}(what, ...)
}
\arguments{
\item{what}{The definition to resolve.}

\item{...}{Additional arguments passed to methods.}

\item{repositories}{A list of \code{\link[=repository]{repositories}} to search in.}

\item{transitive}{A logical indicating if transitive dependencies should be fetched as well}
}
\description{
Resolve objects to file paths
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Tries to resolve the module \code{what} in one of the given \code{repositories}
using \code{\link{fetch_module}}.

\item \code{character}: Checks if \code{what} is a valid path to a file and returns \code{what}.

\item \code{list}: Resolves all elements in \code{what} and returns them as a list.
}}
\examples{
resolve(module('com.h2database:h2:1.3.176'), list(maven_local, maven_central))
}
\seealso{
Other java dependency functions: \code{\link{fetch_module}},
  \code{\link{maven_repository}}, \code{\link{module}}
}

