% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_dependencies.R
\name{resolve_all}
\alias{as_classpath}
\alias{resolve_all}
\title{Resolve a set of module definitions.}
\usage{
resolve_all(..., repositories)

as_classpath(paths)
}
\arguments{
\item{...}{Either local file paths or objects created with \code{module}.}

\item{repositories}{A list of repositories to search in.}

\item{paths}{A character vector of paths, as returned by \code{resolve_all}}
}
\description{
\code{java_classpath} additionally concatenates the resolved paths by the OS path separator.
}

