% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{rjdbc_mapping}
\alias{default_rjdbc_mapping}
\alias{rjdbc_mapping}
\title{Create a converion definition for a set of SQL types between working and transfer type.}
\format{\preformatted{List of 7
 $ :List of 5
  ..$ sql_types   : num [1:3] -6 5 4
  ..$ convert_from:function (x)  
  ..$ convert_to  :function (x)  
  ..$ constraint  :function (x)  
  ..$ create_type : chr "INTEGER"
  ..- attr(*, "class")= chr "RJDBCMapping"
 $ :List of 5
  ..$ sql_types   : num [1:6] 6 7 8 2 3 -5
  ..$ convert_from:function (x)  
  ..$ convert_to  :function (x)  
  ..$ constraint  :function (x)  
  ..$ create_type : chr "DOUBLE PRECISION"
  ..- attr(*, "class")= chr "RJDBCMapping"
 $ :List of 5
  ..$ sql_types   : num [1:2] -7 16
  ..$ convert_from:function (x)  
  ..$ convert_to  :function (x)  
  ..$ constraint  :function (x)  
  ..$ create_type : chr "BOOLEAN"
  ..- attr(*, "class")= chr "RJDBCMapping"
 $ :List of 5
  ..$ sql_types   : num 92
  ..$ convert_from:function (x, ...)  
  ..$ convert_to  :function (x, ...)  
  ..$ constraint  :function (x)  
  ..$ create_type : chr "TIME"
  ..- attr(*, "class")= chr "RJDBCMapping"
 $ :List of 5
  ..$ sql_types   : num 91
  ..$ convert_from:function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 62 5 62 83 5 83 62 62
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ convert_to  :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 63 5 63 51 5 51 63 63
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ constraint  :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 64 5 64 29 5 29 64 64
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ create_type : chr "DATE"
  ..- attr(*, "class")= chr "RJDBCMapping"
 $ :List of 5
  ..$ sql_types   : num 93
  ..$ convert_from:function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 69 5 69 71 5 71 69 69
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ convert_to  :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 70 5 70 36 5 36 70 70
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ constraint  :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 71 5 71 31 5 31 71 71
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ create_type : chr "TIMESTAMP"
  ..- attr(*, "class")= chr "RJDBCMapping"
 $ :List of 5
  ..$ sql_types   : num [1:6] 1 12 -1 -15 -9 -16
  ..$ convert_from:function (x, ...)  
  ..$ convert_to  :function (x, ...)  
  ..$ constraint  :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 78 5 78 47 5 47 78 78
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb68a1dfd78> 
  ..$ create_type : chr "VARCHAR(255)"
  ..- attr(*, "class")= chr "RJDBCMapping"
}}
\usage{
rjdbc_mapping(sql_types, convert_from, convert_to, constraint, create_type)

default_rjdbc_mapping
}
\arguments{
\item{sql_types}{a vector of sql types as defined in \code{\link{JDBC_SQL_TYPES}}}

\item{convert_from}{a function which can be applied to a vector and returns a vector
with the values converted to a desired working type.}

\item{convert_to}{a function which can be applied to a vector of a working type and returns a vector
with values matching to transfer type of the given sql type.}

\item{constraint}{a function which accepts an object argument and returns a logical indicating if this mapping is applicable for the object.}

\item{create_type}{a character which provides the default SQL Type used for creating database columns for objects which satisfy the constraint.}
}
\description{
Create a converion definition for a set of SQL types between working and transfer type.
}
\keyword{datasets}

