% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_mapping.R
\name{write_conversion_rule}
\alias{write_conversion_rule}
\title{Create a write conversion rule}
\usage{
write_conversion_rule(condition, conversion, create_type)
}
\arguments{
\item{condition}{a function which accepts the data and additional attributes and returns a logical
indicating if it is able to convert the data}

\item{conversion}{a function which accepts a data frame column vector and returns it transformed
into a vector of a transfer data type}

\item{create_type}{a character vector of length 1 which holds a SQL type name used to store data
which satisfies the given condition.}
}
\description{
Write conversion rules define how to convert R values to data which can be written to JDBC.
}
\details{
Group conversion rules as a list and pass them to \code{\link{driver}}.
}
\examples{
# Convert difftime vectors into numeric vectors of milliseconds and create 
write_conversion_rule(
  function(data, ...) is.difftime(data),
  function(data, ...) as.numeric(data, units = "secs") * 1000,
  function(...) "TIME"
)
}
\seealso{
Other conversion functions: \code{\link{read_conversion_rule}}
}

