% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{as_classpath}
\alias{as_classpath}
\title{Covert a character vector into a classpath}
\usage{
as_classpath(paths)
}
\arguments{
\item{paths}{A character vector}
}
\value{
A character vector of the concatenated paths.
}
\description{
Concatenate all elements in \code{paths} using \code{.Platform$path.sep} as the separator.
}

