% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JavaUtils.R
\name{checkException}
\alias{checkException}
\title{Check for any pending java exception.}
\usage{
checkException()
}
\value{
\code{NULL} invisibly.
}
\description{
An error is thrown if one is found, otherwise \code{NULL} is returned.
}
\keyword{internal}

