% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCMapping.R
\name{convert_from}
\alias{convert_from}
\title{Convert from transfer type to client type.}
\usage{
convert_from(conversions, data, data_attributes)
}
\arguments{
\item{conversions}{a list of JDBCReadConversion objects}

\item{data}{the data vector to convert}

\item{data_attributes}{a named list of attributes of data}
}
\description{
Convert from transfer type to client type.
}
\keyword{internal}

