% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_to}
\alias{convert_to}
\title{Convert from client type to transfer type}
\usage{
convert_to(conversions, data, data_attributes)
}
\arguments{
\item{conversions}{a list of RJDBCWriteConversion objects}

\item{data}{the data object to convert}

\item{data_attributes}{a named list of attributes}
}
\description{
Convert from client type to transfer type
}

