% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_mapping.R
\name{convert_to_transfer}
\alias{convert_to_transfer}
\title{Convert from client type to transfer type}
\usage{
convert_to_transfer(write_conversion_rules, data, data_attributes)
}
\arguments{
\item{write_conversion_rules}{a list of \code{\link{write_conversion_rule}} objects}

\item{data}{the data object to convert}

\item{data_attributes}{a named list of attributes}
}
\description{
Convert from client type to transfer type
}
\keyword{internal}

