% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCUtils.R
\name{create_j_table}
\alias{create_j_table}
\title{Transform a data frame into a java reference to a info/urbanek/Rpackage/RJDBC/Table}
\usage{
create_j_table(j_statement, data, write_conversions)
}
\arguments{
\item{j_statement}{a jobjRef to a java.sql.PreparedStatement}

\item{data}{a data.frame}

\item{write_conversions}{a list of RJDBCMapping objects}
}
\description{
Transform a data frame into a java reference to a info/urbanek/Rpackage/RJDBC/Table
}

