% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_jdbc_utils.R
\name{create_jdbc_connection}
\alias{create_jdbc_connection}
\title{Establish a JDBC Connection}
\usage{
create_jdbc_connection(url, user, password, ...)
}
\arguments{
\item{url}{the URL of the form \code{jdbc:subprotocol:subname}}

\item{user}{the user to log in}

\item{password}{the user's password}

\item{...}{additional connection arguments}
}
\value{
a jObjRef referencing a java.sql.Connection
}
\description{
Establish a JDBC Connection
}
\keyword{internal}

