% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCUtils.R
\name{create_jdbc_connection}
\alias{create_jdbc_connection}
\title{Create a Java JDBC Connection object}
\usage{
create_jdbc_connection(j_drv, url, user, password, ...)
}
\arguments{
\item{j_drv}{The Java driver object}

\item{url}{the url to connect to}

\item{user}{the user to log in}

\item{password}{the users password}

\item{...}{named values which get transformed into key-value pairs of a 
Java Properties object which is passed to the connect method.}
}
\description{
Create a Java JDBC Connection object
}

