% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCUtils.R
\name{create_jdbc_driver}
\alias{create_jdbc_driver}
\title{Create a Java JDBC Driver object}
\usage{
create_jdbc_driver(driverClass, classPath)
}
\arguments{
\item{driverClass}{the java class name of the JDBC driver to use}

\item{classPath}{a string of paths seperated by \code{path.sep} variable in \code{\link{.Platform}} which should get added to the classpath (see \link[rJava]{.jaddClassPath})}
}
\description{
Create a Java JDBC Driver object
}

