\docType{methods}
\name{dbCallProc,JDBCConnection-method}
\alias{dbCallProc,JDBCConnection-method}
\title{Calls a stored procedure.}
\usage{
\S4method{dbCallProc}{JDBCConnection}(conn, ...)
}
\arguments{
  \item{conn}{a JDBCConnection object}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
Calls a stored procedure.
}

