% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbCallProc,JDBCConnection-method}
\alias{dbCallProc,JDBCConnection-method}
\title{Marked as deprecated in DBI}
\usage{
\S4method{dbCallProc}{JDBCConnection}(conn, name, parameters = list(), ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{name}{the name of the procedure}

\item{parameters}{a list of procedure arguments.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Marked as deprecated in DBI
}

