% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbClearResult,JDBCQueryResult-method}
\alias{dbClearResult,JDBCQueryResult-method}
\title{Clear a result set.}
\usage{
\S4method{dbClearResult}{JDBCQueryResult}(res, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
Clear a result set.
}

