% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbClearResult,JDBCUpdateResult-method}
\alias{dbClearResult,JDBCUpdateResult-method}
\title{Clear a result set.}
\usage{
\S4method{dbClearResult}{JDBCUpdateResult}(res, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCUpdateResult}} object.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
Clear a result set.
}

