% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbCommit,JDBCConnection-method}
\alias{dbCommit,JDBCConnection-method}
\title{Commit a transaction.}
\usage{
\S4method{dbCommit}{JDBCConnection}(conn, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Commit a transaction.
}

