% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbConnect,JDBCConnection-method}
\alias{dbConnect,JDBCConnection-method}
\title{Create a JDBC connection.}
\usage{
\S4method{dbConnect}{JDBCConnection}(drv, password = "", ...)
}
\arguments{
\item{drv}{a \code{\linkS4class{JDBCConnection}} object}

\item{password}{the password for the connection}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Create a connection to the same URL and with the same user as this connection uses.
}

