% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{methods}
\name{dbConnect,JDBCDriver-method}
\alias{dbConnect,JDBCDriver-method}
\title{Create a JDBC connection.}
\usage{
\S4method{dbConnect}{JDBCDriver}(drv, url, user = "", password = "", ...)
}
\arguments{
\item{drv}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{url}{the url to connect to}

\item{user}{the user to log in}

\item{password}{the users password}
}
\description{
Create a JDBC connection.
}

