% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{methods}
\name{dbConnect,JDBCDriver-method}
\alias{dbConnect,JDBCDriver-method}
\title{Connect to a database}
\usage{
\S4method{dbConnect}{JDBCDriver}(drv, url, user = "", password = "",
  sql_dialect = dialect_for_url(url), ...)
}
\arguments{
\item{drv}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{url}{the URL of the form \code{jdbc:subprotocol:subname}}

\item{user}{the user to log in}

\item{password}{the user's password}

\item{sql_dialect}{an \code{\link{sql_dialect}}}

\item{...}{additional connection arguments}
}
\description{
Connect to a database using a
\href{https://docs.oracle.com/javase/tutorial/jdbc/basics/connecting.html#db_connection_url}{JDBC URL}.
}
\examples{
\dontrun{
con <- dbConnect(dbj::driver(), 'jdbc:h2:mem:', 'sa', 'sa')
}
}
\seealso{
Other connection functions: \code{\link{JDBCDriver-class}},
  \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbDisconnect,JDBCConnection-method}}
}

