\docType{methods}
\name{dbDataType,JDBCObject-method}
\alias{dbDataType,JDBCObject-method}
\title{Determine the SQL Data Type of an R object.}
\usage{
\S4method{dbDataType}{JDBCObject}(dbObj, obj, ...)
}
\arguments{
  \item{dbObj}{a \code{\linkS4class{JDBCObject}} object,}

  \item{obj}{an R object whose SQL type we want to
  determine.}

  \item{...}{Needed for compatibility with generic.
  Otherwise ignored.}
}
\description{
Determine the SQL Data Type of an R object.
}

