% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbDisconnect,JDBCConnection-method}
\alias{dbDisconnect,JDBCConnection-method}
\title{Disconnect (close) a connection}
\usage{
\S4method{dbDisconnect}{JDBCConnection}(conn, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Disconnect (close) a connection
}
\seealso{
Other connection functions: \code{\link{JDBCDriver-class}},
  \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbConnect,JDBCDriver-method}}

Other JDBCConnection methods: \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbSendQuery,JDBCConnection,character-method}},
  \code{\link{dbj-edit-tables}},
  \code{\link{dbj-sql-generate}},
  \code{\link{dbj-sql-quote}}, \code{\link{dbj-table}}
}

