% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbDisconnect,JDBCConnection-method}
\alias{dbDisconnect,JDBCConnection-method}
\title{Disconnect an JDBC connection.}
\usage{
\S4method{dbDisconnect}{JDBCConnection}(conn, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Disconnect an JDBC connection.
}

