% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCObjectExtensions.R
\name{dbGetDriver}
\alias{dbGetDriver}
\title{Generics for getting the driver object.}
\usage{
dbGetDriver(dbObj, ...)
}
\arguments{
\item{dbObj}{A \code{JDBCObject} object.}

\item{...}{Other arguments used by methods}
}
\description{
Generics for getting the driver object.
}

