% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection_generics.R
\name{dbGetFields}
\alias{dbGetFields}
\title{Generics for getting a description of table columns available in the specified catalog.}
\usage{
dbGetFields(conn, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object.}

\item{...}{Other arguments used by methods}
}
\description{
Generics for getting a description of table columns available in the specified catalog.
}
\keyword{internal}

