% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection_generics.R, R/JDBCConnection.R
\docType{methods}
\name{dbGetFields}
\alias{dbGetFields}
\alias{dbGetFields,JDBCConnection-method}
\title{Generics for getting a description of table columns available in the specified catalog.}
\usage{
dbGetFields(conn, ...)

\S4method{dbGetFields}{JDBCConnection}(conn, name, pattern = "\%", ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object.}

\item{...}{Other arguments used by methods}

\item{name}{the pattern for table names}

\item{pattern}{the pattern for column names}

\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Generics for getting a description of table columns available in the specified catalog.
}
\keyword{internal}

