% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbGetInfo,JDBCConnection-method}
\alias{dbGetInfo,JDBCConnection-method}
\title{Get info about the connection.}
\usage{
\S4method{dbGetInfo}{JDBCConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of class \code{\linkS4class{JDBCConnection}}}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get info about the connection.
}

