% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R
\docType{methods}
\name{dbGetRowCount,JDBCQueryResult-method}
\alias{dbGetRowCount,JDBCQueryResult-method}
\title{Get number of rows fetched so far}
\usage{
\S4method{dbGetRowCount}{JDBCQueryResult}(res, ...)
}
\description{
Get number of rows fetched so far
}

