% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dbGetRowCount,JDBCUpdateResult-method}
\alias{dbGetRowCount,JDBCUpdateResult-method}
\title{Get number of rows fetched so far}
\usage{
\S4method{dbGetRowCount}{JDBCUpdateResult}(res, ...)
}
\arguments{
\item{res}{
  refers to a \code{DBIResult} object.
}

\item{...}{
  any driver-specific arguments. 
}
}
\description{
Get number of rows fetched so far
}

