% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCUpdateResult.R
\docType{methods}
\name{dbGetRowCount,JDBCUpdateResult-method}
\alias{dbGetRowCount,JDBCUpdateResult-method}
\title{Get number of rows fetched so far}
\usage{
\S4method{dbGetRowCount}{JDBCUpdateResult}(res, ...)
}
\description{
Get number of rows fetched so far
}

