% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnectionExtensions.R, R/JDBCConnection.R
\docType{methods}
\name{dbGetTables}
\alias{dbGetTables}
\alias{dbGetTables,JDBCConnection-method}
\title{Generics for getting a description of the tables available in the given catalog.}
\usage{
dbGetTables(conn, ...)

\S4method{dbGetTables}{JDBCConnection}(conn, pattern = "\%", ...)
}
\arguments{
\item{conn}{An \code{DBIConnection} object.}

\item{...}{Other arguments used by methods}

\item{pattern}{the pattern for table names}
}
\description{
Generics for getting a description of the tables available in the given catalog.
}
\section{Methods (by class)}{
\itemize{
\item \code{JDBCConnection}: Get tables for the JDBCConnection
}}
\keyword{internal}

