\name{dbGetTables}
\alias{dbGetTables}
\title{Generics for getting a description of the tables available in the given catalog.}
\usage{
dbGetTables(conn, ...)
}
\arguments{
  \item{conn}{An \code{DBIConnection} object.}

  \item{...}{Other arguments used by methods}
}
\description{
Generics for getting a description of the tables available
in the given catalog.
}

