% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R
\docType{methods}
\name{dbListFields,JDBCQueryResult,missing-method}
\alias{dbListFields,JDBCQueryResult,missing-method}
\title{List fields in specified table.}
\usage{
\S4method{dbListFields}{JDBCQueryResult,missing}(conn, name, ...)
}
\arguments{
\item{conn}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{name}{Ignored. Needed for compatiblity with generic.}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
List fields in specified table.
}
\keyword{internal}

