% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R
\docType{methods}
\name{dbListFields,JDBCQueryResult,missing-method}
\alias{dbListFields,JDBCQueryResult,missing-method}
\title{Get the names or labels for the columns of the result set.}
\usage{
\S4method{dbListFields}{JDBCQueryResult,missing}(conn, name,
  use_labels = TRUE, ...)
}
\arguments{
\item{conn}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{name}{Ignored. Needed for compatiblity with generic.}

\item{use_labels}{if the the method should return the labels or the names of the columns}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Get the names or labels for the columns of the result set.
}

