% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbListResults,JDBCConnection-method}
\alias{dbListResults,JDBCConnection-method}
\title{List available JDBC result sets.}
\usage{
\S4method{dbListResults}{JDBCConnection}(conn, ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
List available JDBC result sets.
}

