% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbListTables,JDBCConnection-method}
\alias{dbListTables,JDBCConnection-method}
\title{List available JDBC tables.}
\usage{
\S4method{dbListTables}{JDBCConnection}(conn, pattern = "\%", ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{pattern}{the pattern passed to the java method ResultSet.getTables()}

\item{...}{Other parameters passed on to methods.}
}
\description{
List available JDBC tables.
}

