% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbQuoteIdentifier,JDBCConnection,character-method}
\alias{dbQuoteIdentifier,JDBCConnection,character-method}
\title{Quote an identifier using the \code{sql_quote_identifier} function defined in the Drivers \code{sql_dialect} evironment.}
\usage{
\S4method{dbQuoteIdentifier}{JDBCConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{x}{A character vector to label as being escaped SQL.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Quote an identifier using the \code{sql_quote_identifier} function defined in the Drivers \code{sql_dialect} evironment.
}

