% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbQuoteIdentifier,JDBCConnection,SQL-method}
\alias{dbQuoteIdentifier,JDBCConnection,SQL-method}
\title{Quote identifier}
\usage{
\S4method{dbQuoteIdentifier}{JDBCConnection,SQL}(conn, x, ...)
}
\arguments{
\item{conn}{A subclass of DBIConnection, representing an active connection to an DBMS.}

\item{x}{A character vector to label as being escaped SQL.}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Quote identifier
}

