\docType{methods}
\name{dbRollback,JDBCConnection-method}
\alias{dbRollback,JDBCConnection-method}
\title{Roll back a transaction.}
\usage{
\S4method{dbRollback}{JDBCConnection}(conn, ...)
}
\arguments{
  \item{conn}{An existing
  \code{\linkS4class{JDBCConnection}}}

  \item{...}{Ignored. Included for compatibility with
  generic.}
}
\description{
Roll back a transaction.
}

