% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbSendQuery,JDBCConnection,character-method}
\alias{dbSendQuery,JDBCConnection,character-method}
\title{Execute a SQL statement on a database connection}
\usage{
\S4method{dbSendQuery}{JDBCConnection,character}(conn, statement,
  parameters = list(), ...)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{JDBCConnection}}}

\item{statement}{the statement to send over the connection}

\item{parameters}{a list of statment parameters}

\item{...}{Ignored. Needed for compatiblity with generic.}
}
\description{
Execute a SQL statement on a database connection
}

