% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbSendQuery,JDBCConnection,character-method}
\alias{dbSendQuery,JDBCConnection,character-method}
\title{Execute a SQL statement on a database connection}
\usage{
\S4method{dbSendQuery}{JDBCConnection,character}(conn, statement,
  parameters = list(), ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{statement}{the statement to send over the connection}

\item{parameters}{a list of statement parameters}

\item{...}{Other parameters passed on to methods.-JDBCConnection-method}
}
\description{
Execute a SQL statement on a database connection
}

