\name{dbSendUpdate}
\alias{dbSendUpdate}
\title{Generics for sending an update query.}
\usage{
dbSendUpdate(conn, statement, ...)
}
\arguments{
  \item{conn}{An \code{DBIConnection} object.}

  \item{statement}{the stament to send}

  \item{...}{Other arguments used by methods}
}
\description{
Generics for sending an update query.
}

