% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection_generics.R, R/JDBCConnection_update.R
\docType{methods}
\name{dbSendUpdate}
\alias{dbSendUpdate}
\alias{dbSendUpdate,JDBCConnection,character,data.frame-method}
\alias{dbSendUpdate,JDBCConnection,character,list-method}
\alias{dbSendUpdate,JDBCConnection,character,missing-method}
\title{Generics for sending an update query.}
\usage{
dbSendUpdate(conn, statement, parameters, ...)

\S4method{dbSendUpdate}{JDBCConnection,character,missing}(conn, statement,
  parameters, ...)

\S4method{dbSendUpdate}{JDBCConnection,character,list}(conn, statement,
  parameters, ...)

\S4method{dbSendUpdate}{JDBCConnection,character,data.frame}(conn, statement,
  parameters, partition_size = 10000, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object.}

\item{statement}{the statement to send}

\item{parameters}{Optional. Either a named list or a data.frame of statment parameters.
A data.frame will produce a batch update for each row.}

\item{...}{Other arguments used by methods}

\item{partition_size}{the size which will be used to partition the data into seperate commits}

\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{...}{Other parameters passed on to methods.}

\item{statement}{the statement to send over the connection}

\item{parameters}{a list of statement parameters}
}
\description{
Generics for sending an update query.
}
\section{Methods (by class)}{
\itemize{
\item \code{conn = JDBCConnection,statement = character,parameters = missing}: Send update query without parameters

\item \code{conn = JDBCConnection,statement = character,parameters = list}: Send update query with parameters given as a named list

\item \code{conn = JDBCConnection,statement = character,parameters = data.frame}: Send batch update queries with parameters given as a data.frame
}}
\keyword{internal}

