% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCConnection.R, R/JDBCConnectionExtensions.R
\docType{methods}
\name{dbSendUpdate,JDBCConnection,character,missing-method}
\alias{dbSendUpdate}
\alias{dbSendUpdate,JDBCConnection,character,data.frame-method}
\alias{dbSendUpdate,JDBCConnection,character,list-method}
\alias{dbSendUpdate,JDBCConnection,character,missing-method}
\title{Generics for sending an update query.}
\usage{
\S4method{dbSendUpdate}{JDBCConnection,character,missing}(conn, statement,
  parameters, ...)

\S4method{dbSendUpdate}{JDBCConnection,character,list}(conn, statement,
  parameters, ...)

\S4method{dbSendUpdate}{JDBCConnection,character,data.frame}(conn, statement,
  parameters, partition_size = 10000, ...)

dbSendUpdate(conn, statement, parameters, ...)
}
\arguments{
\item{conn}{An \code{DBIConnection} object.}

\item{statement}{the statement to send}

\item{parameters}{Optional. Either a named list or a data.frame of statment parameters.
A data.frame will produce a batch update for each row.}

\item{...}{Other arguments used by methods}

\item{partition_size}{the size which will be used to partition the data into seperate commits}
}
\description{
Generics for sending an update query.
}

