% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dbUnloadDriver,JDBCDriver-method}
\alias{dbUnloadDriver,JDBCDriver-method}
\title{Unload \code{\linkS4class{JDBCDriver}} driver.}
\usage{
\S4method{dbUnloadDriver}{JDBCDriver}(drv, ...)
}
\arguments{
\item{drv}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\value{
A logical indicating whether the operation succeeded or not.
}
\description{
Unload \code{\linkS4class{JDBCDriver}} driver.
}

