% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{methods}
\name{dbWriteTable,JDBCConnection,character,data.frame-method}
\alias{dbWriteTable,JDBCConnection,character,data.frame-method}
\title{Write a local data frame or file to the database.}
\usage{
\S4method{dbWriteTable}{JDBCConnection,character,data.frame}(conn, name, value,
  create = TRUE, append = FALSE, truncate = FALSE, temporary = FALSE,
  row.names = NA, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{name}{character vector of length 1 giving name of table to write to}

\item{value}{the date frame to write to the table}

\item{create}{a logical specifying whether to create a new table if it does not exist. Its default is TRUE.}

\item{append}{a logical specifying whether to append to an existing table. Its default is FALSE}

\item{truncate}{a logical specifying whether to truncate an existing table before appending. Its default is FALSE}

\item{temporary}{\code{TRUE} if the table should be temporary}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

  If \code{TRUE}, always translate row names to a column called "row_names".
  If \code{FALSE}, never translate row names. If \code{NA}, translate
  rownames only if they're a character vector.

  A string is equivalent to \code{TRUE}, but allows you to override the
  default name.

  For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Write a local data frame or file to the database.
}

