% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\name{driver}
\alias{driver}
\title{Factory function for \code{\linkS4class{JDBCDriver}} objects}
\usage{
driver(driverClass, classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions,
  dialect = guess_dialect(driverClass),
  create_new_connection = create_new_dbj_connection)
}
\arguments{
\item{driverClass}{a character vector of length one specifying a JDBC driver class (e.g. 'org.h2.Driver')}

\item{classPath}{a character vector of length one specifying classpaths separated by \code{\link[=.Platform]{path.sep}}
or a character vector of classpaths which will be added to the \code{\link[=.jaddClassPath]{rJava class loader}}}

\item{read_conversions}{a list of \code{\link[=read_conversion]{read conversions}}.}

\item{write_conversions}{a list of \code{\link[=write_conversion]{write conversions}}.}

\item{dialect}{The \code{\link{sql_dialect}} to use.}

\item{create_new_connection}{The factory function for JDBCConnection objects.
Should only be modified by packages extending JDBCDriver and related classes.}
}
\value{
A new \code{\linkS4class{JDBCDriver}}
}
\description{
Call \code{driver} to create a new \code{\linkS4class{JDBCDriver}}
in order to \code{\link[=dbConnect,JDBCDriver-method]{connect}} to databases using the given JDBC driver.
}
\examples{
\dontrun{
drv <- dbj::driver('org.h2.Driver', '~/h2.jar')
con1 <- dbConnect(drv, 'jdbc:h2:mem:')
con2 <- dbConnect(drv, 'jdbc:h2:file:~/foo')
}
}

