% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fetch,JDBCUpdateResult,missing-method}
\alias{fetch,JDBCUpdateResult,missing-method}
\alias{fetch,JDBCUpdateResult,numeric-method}
\title{Fetch records from a previously executed query}
\usage{
\S4method{fetch}{JDBCUpdateResult,missing}(res, n = -1, ...)

\S4method{fetch}{JDBCUpdateResult,numeric}(res, n = -1, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCUpdateResult}} object.}

\item{n}{optional maximum number of records to retrieve per fetch. Use \code{-1} to
retrieve all pending records; use \code{0} for to fetch the default
number of rows as defined in \code{\link{JDBC}}}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\description{
Fetch records from a previously executed query
}

