% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_dependencies.R
\name{fetch_module}
\alias{fetch_module}
\alias{repository}
\title{Fetch a module from a repository}
\usage{
fetch_module(repository, module, ...)
}
\arguments{
\item{repository}{The repository to search in.}

\item{module}{The module to resolve.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The path to the local version of the jar file or NULL if the module is not in the repository.
}
\description{
Classes which support \code{fetch_module} should locate the given \code{module} and make it locally available.
}
\seealso{
Other java dependency functions: \code{\link{maven_repository}},
  \code{\link{module}}, \code{\link{resolve}}
}

