% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection_transactions.R
\docType{methods}
\name{jdbc-transaction}
\alias{dbBegin,JDBCConnection-method}
\alias{dbCommit,JDBCConnection-method}
\alias{dbRollback,JDBCConnection-method}
\alias{jdbc-transaction}
\title{JDBC transaction management.}
\usage{
\S4method{dbBegin}{JDBCConnection}(conn, savepoint_name = NULL)

\S4method{dbCommit}{JDBCConnection}(conn, savepoint_name = NULL)

\S4method{dbRollback}{JDBCConnection}(conn, savepoint_name = NULL)
}
\arguments{
\item{conn}{a \code{\linkS4class{JDBCConnection}} object, produced by
\code{\link[DBI]{dbConnect}}}

\item{savepoint_name}{Supply a name to use a named savepoint. This allows you to
nest multiple transaction}
}
\value{
A boolean, indicating success or failure.
}
\description{
JDBC transaction management.
}

