% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCSQLTypes.R
\docType{data}
\name{JDBC_SQL_TYPES}
\alias{JDBC_SQL_TYPES}
\title{A list of named SQL types as defined in the java class java.sql.Types}
\format{An object of class \code{list} of length 36.}
\usage{
JDBC_SQL_TYPES
}
\description{
A list of named SQL types as defined in the java class java.sql.Types
}
\keyword{datasets}

