% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{JDBC_SQL_TYPES}
\alias{JDBC_SQL_TYPES}
\title{A list of named SQL types as defined in the java class java.sql.Types}
\format{\preformatted{List of 36
 $ BIT          : num -7
 $ TINYINT      : num -6
 $ SMALLINT     : num 5
 $ INTEGER      : num 4
 $ BIGINT       : num -5
 $ FLOAT        : num 6
 $ REAL         : num 7
 $ DOUBLE       : num 8
 $ NUMERIC      : num 2
 $ DECIMAL      : num 3
 $ CHAR         : num 1
 $ VARCHAR      : num 12
 $ LONGVARCHAR  : num -1
 $ DATE         : num 91
 $ TIME         : num 92
 $ TIMESTAMP    : num 93
 $ BINARY       : num -2
 $ VARBINARY    : num -3
 $ LONGVARBINARY: num -4
 $ NULL_        : num 0
 $ OTHER        : num 1111
 $ JAVA_OBJECT  : num 2000
 $ DISTINCT     : num 2001
 $ STRUCT       : num 2002
 $ ARRAY        : num 2003
 $ BLOB         : num 2004
 $ CLOB         : num 2005
 $ REF          : num 2006
 $ DATALINK     : num 70
 $ BOOLEAN      : num 16
 $ ROWID        : num -8
 $ NCHAR        : num -15
 $ NVARCHAR     : num -9
 $ LONGNVARCHAR : num -16
 $ NCLOB        : num 2011
 $ SQLXML       : num 2009
}}
\usage{
JDBC_SQL_TYPES
}
\description{
A list of named SQL types as defined in the java class java.sql.Types
}
\keyword{datasets}

