% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{JDBCConnection-class}
\alias{JDBCConnection}
\alias{JDBCConnection-class}
\title{Class JDBCConnection}
\usage{
JDBCConnection(j_connection, driver)
}
\arguments{
\item{j_connection}{a jobjRef object with a java.sql.Connection reference}

\item{driver}{the full qualified class name of the JDBC Driver to use.}
}
\value{
a \code{\linkS4class{JDBCConnection}} object
}
\description{
Class JDBCConnection

Create a new \code{\linkS4class{JDBCConnection}} object
}

