% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{class}
\name{JDBCConnection-class}
\alias{JDBCConnection}
\alias{JDBCConnection-class}
\alias{SQLKeywords,JDBCConnection-method}
\alias{dbBegin,JDBCConnection-method}
\alias{dbCommit,JDBCConnection-method}
\alias{dbExistsTable,JDBCConnection,character-method}
\alias{dbGetDriver,JDBCConnection-method}
\alias{dbGetException,JDBCConnection-method}
\alias{dbGetInfo,JDBCConnection-method}
\alias{dbListResults,JDBCConnection-method}
\alias{dbReadTable,JDBCConnection,character-method}
\alias{dbRemoveTable,JDBCConnection,character-method}
\alias{dbRollback,JDBCConnection-method}
\alias{dbSQLDialect,JDBCConnection-method}
\alias{show,JDBCConnection-method}
\alias{sqlAppendTable,JDBCConnection-method}
\alias{sqlCreateTable,JDBCConnection-method}
\title{Class JDBCConnection}
\usage{
\S4method{dbExistsTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbReadTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbSQLDialect}{JDBCConnection}(conn)

\S4method{dbBegin}{JDBCConnection}(conn, ...)

\S4method{dbCommit}{JDBCConnection}(conn, ...)

\S4method{dbRollback}{JDBCConnection}(conn, ...)

\S4method{dbGetInfo}{JDBCConnection}(dbObj, ...)

\S4method{dbGetException}{JDBCConnection}(conn, ...)

\S4method{dbListResults}{JDBCConnection}(conn, ...)

\S4method{SQLKeywords}{JDBCConnection}(dbObj, ...)

\S4method{dbGetDriver}{JDBCConnection}(dbObj, ...)

\S4method{show}{JDBCConnection}(object)

\S4method{sqlCreateTable}{JDBCConnection}(con, table, fields, row.names = NA,
  temporary = FALSE, ...)

\S4method{sqlAppendTable}{JDBCConnection}(con, table, values, row.names = NA,
  ...)
}
\arguments{
\item{conn, con, dbObj}{A \code{\linkS4class{JDBCConnection}} object.}

\item{name}{A character string specifying a DBMS table name.}

\item{...}{Arguments passed on to other methods.}

\item{object}{An object of class \code{\linkS4class{JDBCConnection}}}

\item{table}{A table name}

\item{fields}{Either a character vector or a data frame}

\item{row.names}{Either TRUE, FALSE, NA or a string.}

\item{temporary}{If TRUE, will generate a temporary table statement.}

\item{values}{A data frame}
}
\description{
Class JDBCConnection
}
\section{Methods (by generic)}{
\itemize{
\item \code{dbExistsTable}: Does a table exist?

\item \code{dbRemoveTable}: Executes the SQL \code{DROP TABLE}.

\item \code{dbReadTable}: Copy data frames from database.

\item \code{dbSQLDialect}: Get the current dialect

\item \code{dbBegin}: Begin a transaction

\item \code{dbCommit}: Commit a transaction

\item \code{dbRollback}: Rollback a transaction

\item \code{dbGetInfo}: Returns a list with \code{dbname}, \code{db.version}, \code{username}, \code{jdbc_driver_name}, \code{jdbc_driver_version}, \code{url}, \code{host} and \code{port}.

\item \code{dbGetException}: Not implemented. Returns an empty list.

\item \code{dbListResults}: Returns an empty \code{list} as JDBC maintains no list of active results.

\item \code{SQLKeywords}: Returns the list of SQL keywords as defined in the DatabaseMetaData Java object of the associated Java Connection object.

\item \code{dbGetDriver}: Returns the driver for this connection.

\item \code{show}: Prints a short info about the connection

\item \code{sqlCreateTable}: Forwards to sql_create_table in dbSQLDialect

\item \code{sqlAppendTable}: Forwards to sql_create_table in dbSQLDialect
}}
\section{Slots}{

\describe{
\item{\code{j_connection}}{A \code{jobjRef} holding a java.sql.Connection.}

\item{\code{info}}{The connection info list.}

\item{\code{driver}}{A \code{\linkS4class{JDBCDriver}} object.}

\item{\code{create_new_query_result}}{The factory function for \code{\linkS4class{JDBCQueryResult}} objects.}

\item{\code{create_new_update_result}}{The factory function for \code{\linkS4class{JDBCUpdateResult}} objects.}
}}

