% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{class}
\name{JDBCConnection-class}
\alias{JDBCConnection-class}
\alias{SQLKeywords,JDBCConnection-method}
\alias{dbBegin,JDBCConnection-method}
\alias{dbCommit,JDBCConnection-method}
\alias{dbExistsTable,JDBCConnection,character-method}
\alias{dbGetDriver,JDBCConnection-method}
\alias{dbGetException,JDBCConnection-method}
\alias{dbGetInfo,JDBCConnection-method}
\alias{dbListResults,JDBCConnection-method}
\alias{dbReadTable,JDBCConnection,character-method}
\alias{dbRemoveTable,JDBCConnection,character-method}
\alias{dbRollback,JDBCConnection-method}
\title{Class JDBCConnection}
\usage{
\S4method{dbExistsTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbReadTable}{JDBCConnection,character}(conn, name, ...)

\S4method{dbBegin}{JDBCConnection}(conn, ...)

\S4method{dbCommit}{JDBCConnection}(conn, ...)

\S4method{dbRollback}{JDBCConnection}(conn, ...)

\S4method{dbGetInfo}{JDBCConnection}(dbObj, ...)

\S4method{dbGetException}{JDBCConnection}(conn, ...)

\S4method{dbListResults}{JDBCConnection}(conn, ...)

\S4method{SQLKeywords}{JDBCConnection}(dbObj, ...)

\S4method{dbGetDriver}{JDBCConnection}(dbObj, ...)
}
\arguments{
\item{conn}{a \code{\linkS4class{JDBCConnection}} object}

\item{name}{A character string specifying a DBMS table name.}

\item{...}{Arguments passed on to other methods.}

\item{dbObj}{a \code{\linkS4class{JDBCConnection}} object}
}
\description{
Class JDBCConnection
}
\section{Methods (by generic)}{
\itemize{
\item \code{dbExistsTable}: Does a table exist?

\item \code{dbRemoveTable}: Executes the SQL \code{DROP TABLE}.

\item \code{dbReadTable}: Copy data frames from database.

\item \code{dbBegin}: Begin a transaction

\item \code{dbCommit}: Commit a transaction

\item \code{dbRollback}: Rollback a transaction

\item \code{dbGetInfo}: Returns a list with \code{dbname}, \code{db.version}, \code{username}, \code{jdbc_driver_name}, \code{jdbc_driver_version}, \code{url}, \code{identifier_quote_string}, \code{host} and \code{port}.

\item \code{dbGetException}: Not implemented. Returns an empty list.

\item \code{dbListResults}: Returns an empty \code{list} as JDBC maintains no list of active results.

\item \code{SQLKeywords}: Returns the list of SQL keywords as defined in the DatabaseMetaData Java object of the associated Java Connection object.

\item \code{dbGetDriver}: Returns the driver for this connection.
}}

