% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCConnection.R
\docType{class}
\name{JDBCConnection-class}
\alias{JDBCConnection}
\alias{JDBCConnection-class}
\alias{dbGetDriver,JDBCConnection-method}
\alias{dbGetException,JDBCConnection-method}
\alias{dbGetInfo,JDBCConnection-method}
\alias{dbIsValid,JDBCConnection-method}
\alias{dbListResults,JDBCConnection-method}
\alias{dbSQLDialect,JDBCConnection-method}
\alias{show,JDBCConnection-method}
\title{JDBCConnection class}
\usage{
\S4method{dbSQLDialect}{JDBCConnection}(conn)

\S4method{dbGetInfo}{JDBCConnection}(dbObj, ...)

\S4method{dbGetException}{JDBCConnection}(conn, ...)

\S4method{dbListResults}{JDBCConnection}(conn, ...)

\S4method{dbGetDriver}{JDBCConnection}(dbObj, ...)

\S4method{dbIsValid}{JDBCConnection}(dbObj, timeout = 0, ...)

\S4method{show}{JDBCConnection}(object)
}
\arguments{
\item{conn}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{dbObj}{A \code{\linkS4class{JDBCConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{...}{Other parameters passed on to methods.}

\item{timeout}{The time in seconds to wait for the database operation used to validate the connection to complete.
If the timeout period expires before the operation completes, this method returns false.
A value of 0 indicates a timeout is not applied to the database operation.}

\item{object}{An object of class \code{\linkS4class{JDBCConnection}}}
}
\description{
JDBCConnection class
}
\section{Methods (by generic)}{
\itemize{
\item \code{dbSQLDialect}: Get the \code{\link{sql_dialect}} of the driver which created this connection.

\item \code{dbGetInfo}: Returns a named list with the following entries:
    \code{dbname} (the name of the database),
    \code{db.version} (the version of the database),
    \code{username} (the username which was used to connect),
    \code{jdbc_driver_name} (the name of the JDBC driver),
    \code{jdbc_driver_version} (the version of the JDBC driver),
    \code{url} (the url of the JDBC connection),
    \code{host} (currently NULL),
    \code{port} (currently NULL)
and \code{feature.savepoints} (a logical indicating if the database supports savepoints).

\item \code{dbGetException}: Not implemented. Returns an empty list.

\item \code{dbListResults}: Returns an empty list. JDBC does not keep track of active results.

\item \code{dbGetDriver}: Returns the \code{\linkS4class{JDBCDriver}} which created this connection.

\item \code{dbIsValid}: Check if the connection is valid. Returns a logical.

\item \code{show}: Prints a short info about the connection
}}
\section{Slots}{

\describe{
\item{\code{state}}{An environment for mutable states}

\item{\code{j_connection}}{A \code{jobjRef} holding a java.sql.Connection.}

\item{\code{info}}{The connection info list.}

\item{\code{driver}}{A \code{\linkS4class{JDBCDriver}} object.}

\item{\code{create_new_query_result}}{The factory function for \code{\linkS4class{JDBCQueryResult}} objects.}

\item{\code{create_new_update_result}}{The factory function for \code{\linkS4class{JDBCUpdateResult}} objects.}

\item{\code{sql_dialect}}{an \code{\link{sql_dialect}}}
}}
\keyword{internal}

