% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{class}
\name{JDBCDriver-class}
\alias{JDBCDriver-class}
\alias{dbDataType,JDBCDriver-method}
\alias{dbGetDriver,JDBCDriver-method}
\alias{dbGetInfo,JDBCDriver-method}
\alias{dbIsValid,JDBCDriver-method}
\alias{dbListConnections,JDBCDriver-method}
\alias{dbUnloadDriver,JDBCDriver-method}
\alias{driver}
\alias{show,JDBCDriver-method}
\title{Class JDBCDriver with factory methods.}
\usage{
driver(driverClass, classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions,
  dialect = guess_dialect(driverClass))

\S4method{dbListConnections}{JDBCDriver}(drv, ...)

\S4method{dbUnloadDriver}{JDBCDriver}(drv, ...)

\S4method{dbGetInfo}{JDBCDriver}(dbObj, ...)

\S4method{dbGetDriver}{JDBCDriver}(dbObj, ...)

\S4method{dbDataType}{JDBCDriver}(dbObj, obj, ...)

\S4method{dbIsValid}{JDBCDriver}(dbObj, ...)

\S4method{show}{JDBCDriver}(object)
}
\arguments{
\item{driverClass}{the java class name of the JDBC driver to use}

\item{classPath}{a string of paths seperated by \code{path.sep} variable in \code{\link{.Platform}} which should get added to the classpath (see \link[rJava]{.jaddClassPath})}

\item{read_conversions}{a list of JDBCReadConversion objects.}

\item{write_conversions}{a list of JDBCWriteConversion objects.}

\item{dialect}{The \code{\link{sql_dialect}} to use.}

\item{drv}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{...}{Ignored.}

\item{dbObj}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{object}{An object of class \code{\linkS4class{JDBCDriver}}}
}
\description{
\code{JDBCDriver} objects are usually created by 
\code{\link[dbj]{driver}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dbListConnections}: JDBC maintains no list of acitve connections. Returns an empty list.

\item \code{dbUnloadDriver}: Unloading has no effect but returns always \code{TRUE}.

\item \code{dbGetInfo}: Returns a list with \code{driver.version}, \code{client.version} and \code{max.connections}.

\item \code{dbGetDriver}: returns the JDBCDriver object.

\item \code{dbDataType}: returns the DB data type as defined in the \code{write_conversions}.

\item \code{dbIsValid}: always \code{TRUE}.

\item \code{show}: Prints a short info about the driver.
}}

