\docType{class}
\name{JDBCDriver-class}
\alias{JDBC}
\alias{JDBCDriver-class}
\title{Class JDBCDriver with factory method JDBC.}
\usage{
JDBC(driverClass = "", classPath = "", identifier.quote = NA)
}
\arguments{
  \item{driverClass}{the java class name of the JDBC driver
  to use}

  \item{classPath}{a string of paths seperated by : which
  shoul get added to the classpath (see
  \link[rJava]{.jaddClassPath})}

  \item{identifier.quote}{the quoting character to quote
  identifiers}
}
\description{
Class JDBCDriver with factory method JDBC.
}

