% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{class}
\name{JDBCDriver-class}
\alias{JDBCDriver}
\alias{JDBCDriver-class}
\alias{dbDataType,JDBCDriver-method}
\alias{dbGetDriver,JDBCDriver-method}
\alias{dbGetInfo,JDBCDriver-method}
\alias{dbIsValid,JDBCDriver-method}
\alias{dbListConnections,JDBCDriver-method}
\alias{dbUnloadDriver,JDBCDriver-method}
\alias{show,JDBCDriver-method}
\title{JDBCDriver class}
\usage{
\S4method{dbListConnections}{JDBCDriver}(drv, ...)

\S4method{dbUnloadDriver}{JDBCDriver}(drv, ...)

\S4method{dbGetInfo}{JDBCDriver}(dbObj, ...)

\S4method{dbGetDriver}{JDBCDriver}(dbObj, ...)

\S4method{dbDataType}{JDBCDriver}(dbObj, obj, ...)

\S4method{dbIsValid}{JDBCDriver}(dbObj, ...)

\S4method{show}{JDBCDriver}(object)
}
\arguments{
\item{drv}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{...}{Ignored.}

\item{dbObj}{An object of class \code{\linkS4class{JDBCDriver}}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{object}{An object of class \code{\linkS4class{JDBCDriver}}}
}
\description{
\code{JDBCDriver} objects are usually created by 
\code{\link[dbj]{driver}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dbListConnections}: JDBC maintains no list of acitve connections. Returns an empty list.

\item \code{dbUnloadDriver}: Unloading a \code{JDBCDriver} has no effect. Returns \code{TRUE}.

\item \code{dbGetInfo}: Returns a list with \code{driver.version}, \code{client.version} and \code{max.connections}.

\item \code{dbGetDriver}: returns the JDBCDriver object.

\item \code{dbDataType}: returns the DB data type as defined in the \code{write_conversions}.

\item \code{dbIsValid}: always \code{TRUE}.

\item \code{show}: Prints a short info about the driver.
}}
\seealso{
Other connection functions: \code{\link{dbConnect,JDBCConnection-method}},
  \code{\link{dbConnect,JDBCDriver-method}},
  \code{\link{dbDisconnect,JDBCConnection-method}}
}
\keyword{internal}

