% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{class}
\name{JDBCDriver-class}
\alias{JDBC}
\alias{JDBCDriver}
\alias{JDBCDriver-class}
\title{Class JDBCDriver with factory methods.}
\usage{
JDBC(driverClass = "", classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions)

JDBCDriver(driverClass = "", classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions)
}
\arguments{
\item{driverClass}{the java class name of the JDBC driver to use}

\item{classPath}{a string of paths seperated by : which shoul get added to the classpath (see \link[rJava]{.jaddClassPath})}

\item{read_conversions}{a list of RJDBCReadConversion objects.}

\item{write_conversions}{a list of RJDBCWriteConversion objects.}
}
\description{
Class JDBCDriver with factory methods.
}

