% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCDriver.R
\docType{class}
\name{JDBCDriver-class}
\alias{JDBC}
\alias{JDBCDriver}
\alias{JDBCDriver-class}
\alias{dbDataType,JDBCDriver-method}
\alias{dbGetDriver,JDBCDriver-method}
\title{Class JDBCDriver with factory methods.}
\usage{
JDBC(driverClass = "", classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions)

JDBCDriver(driverClass = "", classPath = "",
  read_conversions = default_read_conversions,
  write_conversions = default_write_conversions)

\S4method{dbGetDriver}{JDBCDriver}(dbObj, ...)

\S4method{dbDataType}{JDBCDriver}(dbObj, obj, ...)
}
\arguments{
\item{driverClass}{the java class name of the JDBC driver to use}

\item{classPath}{a string of paths seperated by : which shoul get added to the classpath (see \link[rJava]{.jaddClassPath})}

\item{read_conversions}{a list of JDBCReadConversion objects.}

\item{write_conversions}{a list of JDBCWriteConversion objects.}
}
\description{
\code{JDBCDriver} objects are usually created by
\code{\link[dbj]{JDBC}}.
}
\keyword{internal}

