% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCObject.R
\docType{class}
\name{JDBCObject-class}
\alias{JDBCObject-class}
\alias{SQLKeywords,JDBCObject-method}
\title{Class JDBCObject}
\usage{
\S4method{SQLKeywords}{JDBCObject}(dbObj, ...)
}
\arguments{
\item{dbObj}{an object of type \code{\linkS4class{JDBCObject}}}
}
\description{
Class JDBCObject
}
\section{Methods (by generic)}{
\itemize{
\item \code{SQLKeywords}: Deprecated in DBI
}}
\keyword{internal}

