% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCObject.R
\docType{class}
\name{JDBCObject-class}
\alias{JDBCObject-class}
\alias{SQLKeywords,JDBCObject-method}
\alias{dbDataType,JDBCObject-method}
\title{JDBCObject class}
\usage{
\S4method{SQLKeywords}{JDBCObject}(dbObj, ...)

\S4method{dbDataType}{JDBCObject}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{an object of type \code{\linkS4class{JDBCObject}}}
}
\description{
JDBCObject class
}
\section{Methods (by generic)}{
\itemize{
\item \code{SQLKeywords}: Deprecated in DBI

\item \code{dbDataType}: the method is forwarded to the \code{JDBCDriver} object returnd by \code{dbGetDriver(dbObj)}.
}}
\keyword{internal}

