% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCObject.R
\docType{class}
\name{JDBCObject-class}
\alias{JDBCObject-class}
\alias{SQLKeywords,JDBCObject-method}
\alias{dbDataType,JDBCObject-method}
\alias{dbGetInfo,JDBCObject-method}
\alias{dbIsValid,JDBCObject-method}
\title{Class JDBCObject}
\usage{
\S4method{dbGetInfo}{JDBCObject}(dbObj, ...)

\S4method{dbDataType}{JDBCObject}(dbObj, obj, ...)

\S4method{SQLKeywords}{JDBCObject}(dbObj, ...)

\S4method{dbIsValid}{JDBCObject}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object of class \code{\linkS4class{JDBCObject}}}

\item{...}{Ignored. Included for compatibility with generic.}

\item{obj}{An R object whose SQL type we want to determine.}
}
\description{
Class JDBCObject
}
\keyword{internal}

