% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCQueryResult.R, R/JDBCUpdateResult.R
\docType{class}
\name{JDBCQueryResult-class}
\alias{JDBCQueryResult}
\alias{JDBCQueryResult-class}
\alias{dbClearResult,JDBCQueryResult-method}
\alias{dbClearResult,JDBCUpdateResult-method}
\alias{dbColumnInfo,JDBCUpdateResult-method}
\alias{dbGetInfo,JDBCQueryResult-method}
\alias{dbGetInfo,JDBCUpdateResult-method}
\alias{dbGetRowCount,JDBCQueryResult-method}
\alias{dbGetRowCount,JDBCUpdateResult-method}
\alias{dbGetRowsAffected,JDBCQueryResult-method}
\alias{dbHasCompleted,JDBCQueryResult-method}
\alias{dbIsValid,JDBCQueryResult-method}
\alias{dbIsValid,JDBCUpdateResult-method}
\alias{fetch,JDBCQueryResult,missing-method}
\alias{fetch,JDBCQueryResult,numeric-method}
\alias{fetch,JDBCUpdateResult,missing-method}
\alias{fetch,JDBCUpdateResult,numeric-method}
\title{JDBCQueryResult class.}
\usage{
JDBCQueryResult(j_result_set, connection, statement = "")

\S4method{fetch}{JDBCQueryResult,numeric}(res, n, fetch_size = 0, ...)

\S4method{fetch}{JDBCQueryResult,missing}(res, n = -1, ...)

\S4method{dbClearResult}{JDBCQueryResult}(res, ...)

\S4method{dbGetRowCount}{JDBCQueryResult}(res, ...)

\S4method{dbGetRowsAffected}{JDBCQueryResult}(res, ...)

\S4method{dbHasCompleted}{JDBCQueryResult}(res, ...)

\S4method{dbGetInfo}{JDBCQueryResult}(dbObj, ...)

\S4method{dbIsValid}{JDBCQueryResult}(dbObj, ...)

\S4method{fetch}{JDBCUpdateResult,numeric}(res, n = -1, ...)

\S4method{fetch}{JDBCUpdateResult,missing}(res, n = -1, ...)

\S4method{dbClearResult}{JDBCUpdateResult}(res, ...)

\S4method{dbColumnInfo}{JDBCUpdateResult}(res, ...)

\S4method{dbGetRowCount}{JDBCUpdateResult}(res, ...)

\S4method{dbGetInfo}{JDBCUpdateResult}(dbObj, ...)

\S4method{dbIsValid}{JDBCUpdateResult}(dbObj, ...)
}
\arguments{
\item{j_result_set}{a \code{\linkS4class{jobjRef}} object which holds a reference to a \code{java/sql/ResultSet} Java object.}

\item{connection}{the \code{\linkS4class{JDBCConnection}} which which was used for the query.}

\item{statement}{the stament that was used to create the j_result_set}

\item{res}{an \code{\linkS4class{JDBCQueryResult}} object.}

\item{n}{optional maximum number of records to retrieve per fetch. Use \code{-1} to 
retrieve all pending records; use \code{0} for to fetch the default 
number of rows as defined in \code{\link{JDBC}}}

\item{fetch_size}{a hint to the number of rows that should be fetched from the database in a single block.
See \url{http://docs.oracle.com/javase/7/docs/api/java/sql/Statement.html#setFetchSize(int)}.}

\item{...}{Ignored. Needed for compatibility with generic.}

\item{res}{an \code{\linkS4class{JDBCUpdateResult}} object.}

\item{n}{optional maximum number of records to retrieve per fetch. Use \code{-1} to 
retrieve all pending records; use \code{0} for to fetch the default 
number of rows as defined in \code{\link{JDBC}}}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\value{
a new JDBCQueryResult object
}
\description{
JDBCQueryResult class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fetch}: fetch n results

\item \code{fetch}: fetch all results

\item \code{dbClearResult}: Clear results

\item \code{dbGetRowCount}: Count roes in result set

\item \code{dbGetRowsAffected}: Count rows affected by the last update query

\item \code{dbHasCompleted}: Check if all results have been fetched

\item \code{dbGetInfo}: Get info

\item \code{dbIsValid}: Is the result valid

\item \code{fetch}: Unsupported

\item \code{fetch}: Unsupported

\item \code{dbClearResult}: Has no effect. Returns always \code{TRUE}.

\item \code{dbColumnInfo}: Unsupported

\item \code{dbGetRowCount}: Unsupported

\item \code{dbGetInfo}: Get info

\item \code{dbIsValid}: Is always \code{TRUE}.
}}
\keyword{internal}

