% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{JDBCQueryResult-class}
\alias{JDBCQueryResult}
\alias{JDBCQueryResult-class}
\title{Class JDBCQueryResult with factory method JDBCQueryResult.}
\usage{
JDBCQueryResult(j_result_set, connection, statement = "")
}
\arguments{
\item{j_result_set}{a \code{\linkS4class{jobjRef}} object which holds a reference to a \code{java/sql/ResultSet} Java object.}

\item{connection}{the \code{\linkS4class{JDBCConnection}} which which was used for the query.}

\item{statement}{the stament that was used to create the j_result_set}
}
\value{
a new JDBCQueryResult object
}
\description{
Class JDBCQueryResult with factory method JDBCQueryResult.
}

