% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JDBCUpdateResult.R
\docType{class}
\name{JDBCUpdateResult-class}
\alias{JDBCUpdateResult}
\alias{JDBCUpdateResult-class}
\alias{dbClearResult,JDBCUpdateResult-method}
\alias{dbColumnInfo,JDBCUpdateResult-method}
\alias{dbGetRowCount,JDBCUpdateResult-method}
\title{Class JDBCUpdateResult with factory method JDBCUpdateResult.}
\usage{
JDBCUpdateResult(update_count, connection, statement = "")

\S4method{dbClearResult}{JDBCUpdateResult}(res, ...)

\S4method{dbColumnInfo}{JDBCUpdateResult}(res, ...)

\S4method{dbGetRowCount}{JDBCUpdateResult}(res, ...)
}
\arguments{
\item{update_count}{the number of affected rows.}

\item{connection}{a \code{\linkS4class{JDBCConnection}} object.}

\item{statement}{statement which was used for the query which returned this result}

\item{res}{an \code{\linkS4class{JDBCUpdateResult}} object.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\value{
a new JDBCUpdateResult object
}
\description{
Class JDBCUpdateResult with factory method JDBCUpdateResult.
}
\keyword{internal}

