% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{JDBCUpdateResult-class}
\alias{JDBCUpdateResult}
\alias{JDBCUpdateResult-class}
\title{Class JDBCUpdateResult with factory method JDBCUpdateResult.}
\usage{
JDBCUpdateResult(update_count, connection, statement = "")
}
\arguments{
\item{update_count}{the number of affected rows.}

\item{connection}{a \code{\linkS4class{JDBCConnection}} object.}

\item{statement}{statement which was used for the query which returned this result}
}
\value{
a new JDBCUpdateResult object
}
\description{
Class JDBCUpdateResult with factory method JDBCUpdateResult.
}

