% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCUpdateResult.R
\docType{class}
\name{JDBCUpdateResult-class}
\alias{JDBCUpdateResult}
\alias{JDBCUpdateResult-class}
\alias{dbClearResult,JDBCUpdateResult-method}
\alias{dbColumnInfo,JDBCUpdateResult-method}
\alias{dbGetDriver,JDBCUpdateResult-method}
\alias{dbGetInfo,JDBCUpdateResult-method}
\alias{dbGetRowCount,JDBCUpdateResult-method}
\alias{dbGetRowsAffected,JDBCUpdateResult-method}
\alias{dbGetStatement,JDBCUpdateResult-method}
\alias{dbHasCompleted,JDBCUpdateResult-method}
\alias{dbIsValid,JDBCUpdateResult-method}
\alias{fetch,JDBCUpdateResult,missing-method}
\alias{fetch,JDBCUpdateResult,numeric-method}
\title{Class JDBCUpdateResult with factory method JDBCUpdateResult.}
\usage{
\S4method{fetch}{JDBCUpdateResult,numeric}(res, n = -1, ...)

\S4method{fetch}{JDBCUpdateResult,missing}(res, n = -1, ...)

\S4method{dbClearResult}{JDBCUpdateResult}(res, ...)

\S4method{dbColumnInfo}{JDBCUpdateResult}(res, ...)

\S4method{dbGetRowCount}{JDBCUpdateResult}(res, ...)

\S4method{dbGetInfo}{JDBCUpdateResult}(dbObj, ...)

\S4method{dbIsValid}{JDBCUpdateResult}(dbObj, ...)

\S4method{dbGetDriver}{JDBCUpdateResult}(dbObj, ...)

\S4method{dbHasCompleted}{JDBCUpdateResult}(res, ...)

\S4method{dbGetStatement}{JDBCUpdateResult}(res, ...)

\S4method{dbGetRowsAffected}{JDBCUpdateResult}(res, ...)
}
\arguments{
\item{res}{an \code{\linkS4class{JDBCUpdateResult}} object.}

\item{n}{optional maximum number of records to retrieve per fetch. Use \code{-1} to 
retrieve all pending records; use \code{0} for to fetch the default 
number of rows as defined in \code{\link{JDBC}}}

\item{...}{Ignored. Needed for compatibility with generic.}

\item{dbObj}{An object of class \code{\linkS4class{JDBCUpdateResult}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Class JDBCUpdateResult with factory method JDBCUpdateResult.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fetch}: Returns an empty data frame.

\item \code{fetch}: Returns an empty data frame.

\item \code{dbClearResult}: Has no effect. Returns always \code{TRUE}.

\item \code{dbColumnInfo}: Unsupported

\item \code{dbGetRowCount}: Unsupported

\item \code{dbGetInfo}: Get info

\item \code{dbIsValid}: Is always \code{TRUE}.
}}
\section{Methods}{

\code{dbHasCompleted}: Check if all results have been fetched


\code{dbGetStatement}: Returns the statement that was passed to dbSendQuery


\code{dbGetRowsAffected}: Returns the number of rows that were added, deleted, or updated
}
\keyword{internal}

