% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_dependencies.R
\docType{data}
\name{maven_repository}
\alias{maven_central}
\alias{maven_local}
\alias{maven_local_repository}
\alias{maven_remote_repository}
\alias{maven_repository}
\title{Maven repository}
\format{An object of class \code{maven_remote_repository} (inherits from \code{maven_repository}, \code{module_repository}) of length 2.}
\usage{
maven_remote_repository(url, install = TRUE)

maven_local_repository(path)

maven_central

maven_local
}
\arguments{
\item{url, path}{The location of the reopsitory.}

\item{install}{Should the fetched artifact be installed to the local repository?}
}
\description{
Use maven repositories to locate modules.
}
\keyword{datasets}

