% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven.R
\name{module}
\alias{module}
\alias{module.character}
\alias{module.list}
\title{Generic to create a module definition.}
\usage{
module(x)

\method{module}{character}(x)

\method{module}{list}(x)
}
\arguments{
\item{x}{The definition.}
}
\description{
Generic to create a module definition.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Construct from a string like 'group:name:version'.

\item \code{list}: Construct from a list with character elements named \code{group}, \code{name} and \code{version}.
}}

