% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{partition}
\alias{partition}
\alias{partition.data.frame}
\alias{partition.default}
\title{Partition data into chunks}
\usage{
partition(data, size)

\method{partition}{data.frame}(data, size)

\method{partition}{default}(data, size)
}
\arguments{
\item{data}{the data to partition}

\item{size}{the partition size}
}
\description{
Partition a vector, list matrix or data.frame into a list of vectors,
lists, matrices or data.frames respectively,
each of the same size (last might be smaller).
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Partition a data.frame

\item \code{default}: Default partition function
}}

