% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_mapping.R
\docType{data}
\name{read_conversion}
\alias{default_read_conversions}
\alias{read_conversion}
\alias{sqltype_read_conversion}
\title{Create a JDBCReadConversion object.}
\format{The default conversions}
\usage{
read_conversion(condition, r_class, conversion)

sqltype_read_conversion(sql_types, r_class, conversion)

default_read_conversions
}
\arguments{
\item{condition}{a function which accepts a list and returns a logical}

\item{r_class}{the target class of the conversion}

\item{conversion}{a function which accepts a data frame column vector
and returns it transformed into a vector of a transfer data type.}

\item{sql_types}{a numeric vector of JDBC_SQL_TYPES}
}
\description{
Create a JDBCReadConversion object.

The dafault read conversions
}
\seealso{
Other conversion.functions: \code{\link{write_conversion}}
}
\keyword{datasets}

