% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JDBCMapping.R
\docType{data}
\name{read_conversion}
\alias{default_read_conversions}
\alias{read_conversion}
\alias{sqltype_read_conversion}
\title{Create a RJDBCReadConversion object.}
\format{\preformatted{List of 7
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (x)  
  ..- attr(*, "class")= chr "RJDBCReadConversion"
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (x)  
  ..- attr(*, "class")= chr "RJDBCReadConversion"
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (x)  
  ..- attr(*, "class")= chr "RJDBCReadConversion"
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (x)  
  ..- attr(*, "class")= chr "RJDBCReadConversion"
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (x)  
  ..- attr(*, "class")= chr "RJDBCReadConversion"
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (data)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 57 5 57 89 5 89 57 57
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..- attr(*, "class")= chr "RJDBCReadConversion"
 $ :List of 2
  ..$ condition :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 19 29 3 19 3 26 29
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..$ conversion:function (data)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 61 5 61 77 5 77 61 61
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f862e69f978> 
  ..- attr(*, "class")= chr "RJDBCReadConversion"
}}
\usage{
read_conversion(condition, conversion)

sqltype_read_conversion(sql_types, conversion)

default_read_conversions
}
\arguments{
\item{condition}{a function which accepts a list and returns a logical}

\item{conversion}{a function which accepts a data frame column vector and returns it transformed into a vector of a transfer data type.}

\item{sql_types}{a numeric vector of JDBC_SQL_TYPES}
}
\description{
Create a RJDBCReadConversion object.
}
\keyword{datasets}

