% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maven.R
\name{resolve}
\alias{resolve}
\alias{resolve.maven_repository}
\title{Resolve a module in a given repository.}
\usage{
resolve(repository, module, ...)

\method{resolve}{maven_repository}(repository, module, ...)
}
\arguments{
\item{repository}{The repository to search in.}

\item{module}{The module to resolve.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Resolve a module in a given repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{maven_repository}: Resove in a \code{maven_repository}.
}}

