% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_dependencies.R
\name{resolve}
\alias{resolve}
\alias{resolve.character}
\alias{resolve.list}
\alias{resolve.module}
\title{Generic to resolve a module or file.}
\usage{
resolve(what, ...)

\method{resolve}{module}(what, repositories, ...)

\method{resolve}{character}(what, ...)

\method{resolve}{list}(what, ...)
}
\arguments{
\item{what}{The definition to resolve.}

\item{...}{Additional arguments passed to methods.}

\item{repositories}{A list of repositories to search in.}
}
\description{
Generic to resolve a module or file.
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Tries to resolve the module \code{what} in one of the given \code{repositories},
installs it, and returns the local path.

\item \code{character}: Checks if \code{what} is a valid path to a file and returns \code{what}.

\item \code{list}: Resolves all elements in \code{what} and returns them as a list.
}}

