% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLDialect.R
\docType{data}
\name{sql_dialect}
\alias{generic_append_table}
\alias{generic_clear_table}
\alias{generic_create_table}
\alias{generic_sql}
\alias{sql_dialect}
\title{Create a new SQL dialect environment}
\format{An object of class \code{sql_dialect} of length 2.}
\usage{
generic_create_table(conn, table, data, temporary = FALSE)

generic_append_table(conn, table, data)

generic_clear_table(conn, table, use_delete = FALSE)

sql_dialect(name, sql_create_table = generic_create_table,
  sql_append_table = generic_append_table,
  sql_clear_table = generic_clear_table)

generic_sql
}
\arguments{
\item{conn}{An object of type \code{\linkS4class{JDBCConnection}}}

\item{table}{The table name}

\item{data}{A data.frame}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}

\item{use_delete}{If \code{TRUE}, will use DELETE. If \code{FALSE}, TRUNCATE.}

\item{name}{The name of the dialect}

\item{sql_create_table}{A function which generates an SQL statement for creating a table}

\item{sql_append_table}{A function which generates an SQL statement for adding data to the table}

\item{sql_clear_table}{A function which generates an SQL statement for truncating a table}
}
\value{
A new structure with class \code{sql_dialect}.
}
\description{
Create a new SQL dialect environment
}
\keyword{datasets}

