% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLDialect.R
\docType{data}
\name{sql_dialect}
\alias{generic_append_table}
\alias{generic_clear_table}
\alias{generic_create_table}
\alias{generic_quote_identifier}
\alias{generic_quote_string}
\alias{generic_sql}
\alias{guess_dialect}
\alias{sql_dialect}
\title{Create a new SQL dialect environment}
\format{An object of class \code{sql_dialect} of length 2.}
\usage{
generic_create_table(conn, table, data, temporary = FALSE)

generic_append_table(conn, table, data)

generic_clear_table(conn, table, use_delete = FALSE)

generic_quote_identifier(conn, x, ...)

generic_quote_string(conn, x, ...)

sql_dialect(name, sql_create_table = generic_create_table,
  sql_append_table = generic_append_table,
  sql_clear_table = generic_clear_table,
  sql_quote_identifier = generic_quote_identifier,
  sql_quote_string = generic_quote_string)

generic_sql

guess_dialect(driver_class)
}
\arguments{
\item{conn}{An object of type \code{\linkS4class{JDBCConnection}}}

\item{table}{The table name}

\item{data}{A data.frame}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}

\item{use_delete}{If \code{TRUE}, will use DELETE. If \code{FALSE}, TRUNCATE.}

\item{x}{A character vector to label as being escaped SQL.}

\item{...}{Other parameters passed on to methods.}

\item{name}{The name of the dialect}

\item{sql_create_table}{A function which generates an SQL statement for creating a table}

\item{sql_append_table}{A function which generates an SQL statement for adding data to the table}

\item{sql_clear_table}{A function which generates an SQL statement for truncating a table}

\item{sql_quote_identifier}{The function called by the \code{dbQuoteIdentifier,JDBCConnection-character-method} method}

\item{sql_quote_string}{The function called by the \code{dbQuoteString,JDBCConnection-character-method} method}

\item{driver_class}{The full classname of a Java Driver class.}
}
\value{
A new structure with class \code{sql_dialect}.
}
\description{
Create a new SQL dialect environment
}
\keyword{datasets}

