% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sql_escape}
\alias{sql_escape}
\alias{sql_escape_identifier}
\alias{sql_escape_value}
\title{Escape a string so that is is safe to use inside a SQL query.}
\usage{
sql_escape(string, identifier = FALSE, quote = "\\"")

sql_escape_identifier(string, quote = "\\"")

sql_escape_value(string, quote = "\\"")
}
\arguments{
\item{string}{a character vector}

\item{identifier}{a logical value idicating if \code{string} is an identifier}

\item{quote}{the quoting character}
}
\description{
Escape a string so that is is safe to use inside a SQL query.
}

