% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLUtils.R
\name{sql_escape}
\alias{sql_escape}
\alias{sql_escape_identifier}
\alias{sql_escape_value}
\title{Escape a string so that is is safe to use inside a SQL query.}
\usage{
sql_escape(string, quote, identifier)

sql_escape_identifier(string, quote)

sql_escape_value(string, quote)
}
\arguments{
\item{string}{a character vector}

\item{quote}{the quoting character}

\item{identifier}{a logical value idicating if \code{string} is an identifier}
}
\description{
Escape a string so that is is safe to use inside a SQL query.

Escape a SQL identifier

Escape a SQL value
}

