% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JavaUtils.R
\name{verifyNotNull}
\alias{verifyNotNull}
\title{Verify that the given \code{j_object} is an object of type \code{jobjRef} but not \code{.jnull()}.}
\usage{
verifyNotNull(j_object, ...)
}
\arguments{
\item{j_object}{an object of type \code{\linkS4class{jobjRef}}}

\item{...}{any further message parts passed to \code{stop}}
}
\value{
\code{NULL} invisibly.
}
\description{
An error is thrown if the verification fails, otherwise \code{NULL} is returned.
}
\keyword{internal}

