% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDBCMapping.R
\docType{data}
\name{write_conversion}
\alias{default_write_conversions}
\alias{mapped_write_conversion}
\alias{write_conversion}
\title{Create a JDBCWriteConversion object.}
\format{The default conversions}
\usage{
write_conversion(condition, conversion, create_type)

mapped_write_conversion(class_names, conversion, create_type)

default_write_conversions
}
\arguments{
\item{condition}{a function which accepts a list and returns a logical.}

\item{conversion}{a function which accepts a data frame column vector and returns it transformed into a vector of a transfer data type.}

\item{create_type}{a character vector of length 1 which holds a sql type name used to store data which sattisfies the given condition.}

\item{class_names}{a character vector of class names}
}
\description{
Create a JDBCWriteConversion object.
}
\seealso{
Other conversion.functions: \code{\link{read_conversion}}
}
\keyword{datasets}

