% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{write_conversion}
\alias{default_write_conversions}
\alias{mapped_write_conversion}
\alias{write_conversion}
\title{Create a RJDBCWriteConversion object.}
\format{\preformatted{List of 7
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (x, ...)  
  ..$ create_type: chr "INTEGER"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (x)  
  ..$ create_type: chr "DOUBLE PRECISION"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (x, ...)  
  ..$ create_type: chr "BOOLEAN"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (x, ...)  
  ..$ create_type: chr "TIME"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (data)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 121 5 121 57 5 57 121 121
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ create_type: chr "DATE"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (data)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 126 5 126 42 5 42 126 126
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ create_type: chr "TIMESTAMP"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
 $ :List of 3
  ..$ condition  :function (attributes)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 20 93 3 20 3 90 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11f6aa200> 
  ..$ conversion :function (x)  
  ..$ create_type: chr "VARCHAR(255)"
  ..- attr(*, "class")= chr "RJDBCWriteConversion"
}}
\usage{
write_conversion(condition, conversion, create_type)

mapped_write_conversion(class_names, conversion, create_type)

'default_write_conversions'
}
\arguments{
\item{condition}{a function which accepts a list and returns a logical.}

\item{conversion}{a function which accepts a data frame column vector and returns it transformed into a vector of a transfer data type.}

\item{create_type}{a character vector of length 1 which holds a sql type name used to store data which sattisfies the given condition.}

\item{class_names}{a character vector of class names}
}
\description{
Create a RJDBCWriteConversion object.
}
\keyword{datasets}

